% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ancestries}
\alias{ancestries}
\title{Plot simulated ancestry proportions}
\usage{
ancestries(model, generation_time = FALSE)
}
\arguments{
\item{model}{Compiled model object}
}
\description{
Plot simulated ancestry proportions
}
