% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{animate}
\alias{animate}
\title{Animate the simulated population dynamics}
\usage{
animate(
  model,
  output_dir = model$path,
  output_prefix = "output",
  steps,
  gif = NULL,
  width = 800,
  height = 560
)
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}

\item{output_dir}{A directory where to look for simulation outputs (by default,
all output files are in a model directory)}

\item{output_prefix}{A common prefix of output files (by default, all files
will share a prefix \code{"output"})}

\item{steps}{How many frames should the animation have?}

\item{gif}{Path to an output GIF file (animation object returned
by default)}
}
\value{
If \code{gif = NULL}, return gganimate animation object. Otherwise a GIF
file is saved and no value is returned.
}
\description{
Animate the simulated population dynamics
}
