% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{boundary}
\alias{boundary}
\title{Deprecated function}
\usage{
boundary(pop, time, center = NULL, radius = NULL, polygon = NULL, lock = FALSE)
}
\arguments{
\item{pop}{Object of the class \code{slendr_pop}}

\item{time}{Time of the change}

\item{center}{Two-dimensional vector specifying the center of the
circular range}

\item{radius}{Radius of the circular range}

\item{polygon}{List of vector pairs, defining corners of the
polygon range (see also the \code{region} argument) or a
geographic region of the class \code{slendr_region} from which
the polygon coordinates will be extracted}

\item{lock}{Maintain the same density of individuals. If
\code{FALSE} (the default), the number of individuals in the
population will not change. If \code{TRUE}, the number of
individuals simulated will be changed (increased or decreased)
appropriately, to match the new population range area.}
}
\description{
Deprecated function
}
