% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ranges.R
\name{check_not_intersected}
\alias{check_not_intersected}
\title{Check whether given population region has not yet been intersected}
\usage{
check_not_intersected(pop)
}
\description{
Check whether given population region has not yet been intersected
}
