% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{check_not_rendered}
\alias{check_not_rendered}
\title{Check whether given population region has not yet been rendered}
\usage{
check_not_rendered(pop)
}
\arguments{
\item{pop}{Spatial population object of the 'spammr_pop' class}
}
\description{
Check whether given population region has not yet been rendered
}
