% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_resolution}
\alias{check_resolution}
\title{Check whether the given value makes sense given the map dimensions}
\usage{
check_resolution(map, val)
}
\description{
Check whether the given value makes sense given the map dimensions
}
