% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_split_time}
\alias{check_split_time}
\title{Check the consistency of the given split time to the parent population}
\usage{
check_split_time(time, parent)
}
\description{
Check the consistency of the given split time to the parent population
}
