% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{compile}
\alias{compile}
\title{Deprecated function}
\usage{
compile(
  populations,
  generation_time,
  path = NULL,
  resolution = NULL,
  competition = NULL,
  mating = NULL,
  dispersal = NULL,
  geneflow = list(),
  overwrite = FALSE,
  sim_length = NULL,
  direction = NULL,
  slim_script = system.file("scripts", "script.slim", package = "slendr"),
  description = "",
  dir = NULL
)
}
\arguments{
\item{populations}{Object(s) of the \code{slendr_pop} class (multiple objects
need to be specified in a \code{list})}

\item{generation_time}{Generation time (in model time units)}

\item{path}{Output directory for the model configuration files which will be
loaded by the backend SLiM script. If \code{NULL}, model configuration
files will be saved to a temporary directory.}

\item{resolution}{How many distance units per pixel?}

\item{competition}{Maximum spatial competition and mating choice
distance}

\item{mating}{Maximum spatial competition and mating choice
distance}

\item{dispersal}{Standard deviation of the normal distribution of the
parent-offspring distance}

\item{geneflow}{Deprecated. Use \code{gene_flow} instead.}

\item{overwrite}{Completely delete the specified directory, in case it
already exists, and create a new one?}

\item{sim_length}{Total length of the simulation (required for forward time
models, optional for models specified in backward time units which by
default run to "the present time")}

\item{direction}{Intended direction of time. Under normal circumstances this
parameter is inferred from the model and does not need to be set manually.}

\item{slim_script}{Path to a SLiM script to be used for executing the model
(by default, a bundled backend script will be used)}

\item{description}{Optional short description of the model}

\item{dir}{Deprecated. Use \code{path} instead.}
}
\description{
Deprecated function
}
