% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demes.R
\name{compile_demes}
\alias{compile_demes}
\title{Load and validate a Demes model in slendr format}
\usage{
compile_demes(demes_path, output_path = NULL, overwrite = FALSE)
}
\arguments{
\item{demes_path}{File path to yaml file specifying the Demes model}

\item{output_path}{File path where a folder with the resulting slendr model will be saved.
By default it is NULL and no model is saved to disk.}

\item{overwrite}{If the output folder already exists, should it be overwritten? Defaults to FALSE.}
}
\value{
Compiled \code{slendr_model} model object which encapsulates all
information about the specified model (which populations are involved,
when and how much gene flow should occur, what is the spatial resolution
of a map, and what spatial dispersal and mating parameters should be used
in a SLiM simulation, if applicable)
}
\description{
Load and validate a Demes model in slendr format
}
\examples{
path_jacobs <- system.file("extdata", "models", "Demes", "jacobs.yaml", package = "slendr")
slendr_model <- compile_demes(path_jacobs)
}
