% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{compile_maps}
\alias{compile_maps}
\title{Process vectorized population boundaries into a table with
rasterized map objects}
\usage{
compile_maps(populations, splits_table, pixel)
}
\description{
Process vectorized population boundaries into a table with
rasterized map objects
}
