% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{compile_splits}
\alias{compile_splits}
\title{Iterate over population objects and convert he information about
population split hierarchy and split times into a data frame}
\usage{
compile_splits(populations)
}
\description{
Iterate over population objects and convert he information about
population split hierarchy and split times into a data frame
}
