% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{convert}
\alias{convert}
\title{Project a location of a point (latitude, longitude) on a raster}
\usage{
convert(lat, lon, model)
}
\arguments{
\item{lat, lon}{Geographic system coordinates}

\item{model}{Model object}
}
\value{
Two-dimensional numeric vector of pixel coordinates
}
\description{
Project a location of a point (latitude, longitude) on a raster
}
