% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_time}
\alias{convert_time}
\title{Convert time from backward to forward direction}
\usage{
convert_time(df, direction, columns, max_time, generation_time)
}
\description{
Convert time from backward to forward direction
}
