% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_attributes}
\alias{copy_attributes}
\title{Transfer given set of attributes from one object to another}
\usage{
copy_attributes(to, from, which)
}
\description{
Transfer given set of attributes from one object to another
}
