% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{create_polygon}
\alias{create_polygon}
\title{Create a simple geometry polygon object from the list of
coordinates}
\usage{
create_polygon(coords, world)
}
\arguments{
\item{coords}{List of vectors (pairs of longitude and latitude values)}

\item{world}{Object of the type 'sf' which defines the world}
}
\description{
Create a simple geometry polygon object from the list of
coordinates
}
