% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{define_boundary}
\alias{define_boundary}
\title{Define a range (simple geometry object) for a population or a
geographic region}
\usage{
define_boundary(map, center = NULL, radius = NULL, coords = NULL)
}
\description{
Define a range (simple geometry object) for a population or a
geographic region
}
