% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{define_zoom}
\alias{define_zoom}
\title{Define a rectangular region for zooming in on a part of the world}
\usage{
define_zoom(lon, lat, source_crs = "EPSG:4326")

define_zoom(lon, lat, source_crs = "EPSG:4326")
}
\description{
Define a rectangular region for zooming in on a part of the world

Define a rectangular region for zooming in on a part of the world
}
