% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{distance}
\alias{distance}
\title{Calculate the distance between a pair of spatial boundaries}
\usage{
distance(x, y, measure, time = NULL)
}
\arguments{
\item{x, y}{Objects of the class \code{slendr}}

\item{time}{Time closest to the spatial maps of \code{x} and \code{y} if they
represent \code{slendr_pop} population boundaries (ignored for general
\code{slendr_region} objects)}
}
\value{
If the coordinate reference system was specified, a distance in
projected units (i.e. meters) is returned. Otherwise the function returns a
normal Euclidian distance.
}
\description{
Calculate the distance between a pair of spatial boundaries
}
