% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{expand}
\alias{expand}
\title{Expand population radius by a given factor in a given time}
\usage{
expand(pop, by, duration, snapshots, region = NULL)
}
\arguments{
\item{pop}{Spatial population object of \code{spammr_pop} class}

\item{by}{How many kilometers to expand by?}

\item{duration}{Duration of the spatial population expansion}

\item{snapshots}{Number of time slices to split the movement into}

\item{region}{Geographic region to restrict the expansion to}
}
\value{
Object of the \code{spammr_pop} (and \code{sf}) class
}
\description{
Expand population radius by a given factor in a given time
}
