% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{explore_model}
\alias{explore_model}
\title{Open an interactive browser of the spatial model}
\usage{
explore_model(model)
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}
}
\value{
No return value, called in order to start an interactive browser-based
interface to explore the dynamics of a slendr model
}
\description{
Open an interactive browser of the spatial model
}
