% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_admixture_edges}
\alias{get_admixture_edges}
\title{Create a table of population admixture edges for graph visualization}
\usage{
get_admixture_edges(admix_table)
}
\description{
Create a table of population admixture edges for graph visualization
}
