% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_env}
\alias{get_env}
\title{Get the name of the current slendr Python environment}
\usage{
get_env()
}
\value{
Name of the slendr Python environment
}
\description{
Get the name of the current slendr Python environment
}
