% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_geneflow_edges}
\alias{get_geneflow_edges}
\title{Create a table of population geneflow edges for graph visualization}
\usage{
get_geneflow_edges(admix_table)
}
\description{
Create a table of population geneflow edges for graph visualization
}
