% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geneflows}
\alias{get_geneflows}
\title{Get a data frame of geneflow events active at a given time point}
\usage{
get_geneflows(model, time)
}
\description{
Get a data frame of geneflow events active at a given time point
}
