% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_intermediate_edges}
\alias{get_intermediate_edges}
\title{Create a table of 'intermediate' edges for graph visualization}
\usage{
get_intermediate_edges(split_edges, admix_edges)
}
\description{
Create a table of 'intermediate' edges for graph visualization
}
