% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lineage_splits}
\alias{get_lineage_splits}
\title{Get split times of populations in the lineage of the given population}
\usage{
get_lineage_splits(x)
}
\description{
Get split times of populations in the lineage of the given population
}
