% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_migrations}
\alias{get_migrations}
\title{Get a data frame of migration events active at a given time point}
\usage{
get_migrations(model, time)
}
\description{
Get a data frame of migration events active at a given time point
}
