% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_python}
\alias{get_python}
\title{Get a path to internal Python interpreter of slendr}
\usage{
get_python()
}
\value{
A character scalar path to slendr's Python binary
}
\description{
Get a path to internal Python interpreter of slendr
}
