% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_split_edges}
\alias{get_split_edges}
\title{Create a table of population split edges for graph visualization}
\usage{
get_split_edges(split_table)
}
\description{
Create a table of population split edges for graph visualization
}
