% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_terminal_edges}
\alias{get_terminal_edges}
\title{Generate edges leading to the times of population removals in SLiM}
\usage{
get_terminal_edges(split_edges, geneflow_edges, split_table)
}
\description{
Generate edges leading to the times of population removals in SLiM
}
