% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{intersect_features}
\alias{intersect_features}
\title{Take a population region and intersect them with the
underlying world map or its geographic region}
\usage{
intersect_features(pop)
}
\description{
Take a population region and intersect them with the
underlying world map or its geographic region
}
