% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{load}
\alias{load}
\title{Load a previously serialized spammr model}
\usage{
load(model_dir)
}
\arguments{
\item{model_dir}{Directory with all required configuration files}
}
\description{
Load a previously serialized spammr model
}
