% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{map}
\alias{map}
\title{Define a world map for all spatial operations}
\usage{
map(xrange, yrange, crs = "EPSG:4326")
}
\arguments{
\item{xrange}{Numeric vector with minimum and maximum longitude}

\item{yrange}{Numeric vector with minimum and maximum latitude}

\item{crs}{Coordinate Reference System to use for all spatial
operations (default is WGS-84 or EPSG:4326 CRS)}
}
\value{
Object of the \code{spammr_world} (and \code{sf}) class
}
\description{
Download Natural Earth land area map data, zoom on a given window
determined by longitude and latitude coordinates and transform to a
specified projection
}
