% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{move}
\alias{move}
\title{Move population to a new location in a given amount of time}
\usage{
move(pop, trajectory, end, snapshots, start = NULL)
}
\arguments{
\item{pop}{Spatial population object of \code{spannr_pop} class}

\item{trajectory}{List of two-dimensional vectors \link{(longitude, latitude)}
specifying the trajectory of the population movement}

\item{snapshots}{Number of time slices to split the movement into}

\item{start, end}{Start/end points of the population movement}
}
\value{
Object of the \code{spannr_pop} (and \code{sf}) class
}
\description{
Move population to a new location in a given amount of time
}
