% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{overlap}
\alias{overlap}
\title{Generate the overlap of two \code{slendr} objects}
\usage{
overlap(x, y, name = NULL)
}
\arguments{
\item{x}{Object of the class \code{slendr}}

\item{y}{Object of the class \code{slendr}}

\item{name}{Optional name of the resulting geographic region. If missing,
name will be constructed from the function arguments.}
}
\value{
Object of the class \code{slendr_region}
}
\description{
Generate the overlap of two \code{slendr} objects
}
