% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{plot.slendr}
\alias{plot.slendr}
\title{Deprecated function}
\usage{
\method{plot}{slendr}(
  ...,
  time = NULL,
  geneflows = FALSE,
  graticules = "original",
  intersect = TRUE,
  show_map = TRUE,
  title = NULL,
  interpolated_maps = NULL
)
}
\description{
Deprecated function
}
