% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot.spamr}
\alias{plot.spamr}
\title{Plot spatio-temporal population distributions}
\usage{
\method{plot}{spamr}(..., snapshots = F, rendering = F)
}
\arguments{
\item{...}{Population/geographic region objects of the 'spamr'
class}

\item{snapshots}{Plot time snapshots in individual panels?}
}
\description{
Plot the spatio-temporal distributions of populations and
geographic regions across a map.
}
\details{
If only geographic regions are given, they are colored. If both
them and populations are given, only populations are specified.
}
