% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spannr.R
\name{plot.spannr}
\alias{plot.spannr}
\title{Plot \code{spannr} geographic features on a map}
\usage{
\method{plot}{spannr}(
  ...,
  pop_facets = TRUE,
  time_facets = FALSE,
  intersect = TRUE,
  graticules = "original",
  title = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{...}{Objects of classes \code{spannr_map},
\code{spannr_region}, or \code{spannr_pop}}

\item{pop_facets}{Plot populations in individual panels?}

\item{time_facets}{Plot time snapshots in individual panels?}

\item{intersect}{Intersect the population boundaries against
landscape and other geographic boundaries (default TRUE)?}

\item{graticules}{Plot graticules in the original Coordinate
Reference System (such as longitude-latitude), or in the internal
CRS (such as meters)?}

\item{title}{Plot title}

\item{nrow, ncol}{Number of columns or rows in the facet plot}
}
\description{
Plots objects of the three \code{spannr} spatial classes
(\code{spannr_map}, \code{spannr_region}, and \code{spannr_pop}).
}
