% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{plot.stamp}
\alias{plot.stamp}
\title{Plot spatio-temporal population distributions}
\usage{
\method{plot}{stamp}(
  ...,
  pop_facets = TRUE,
  time_facets = FALSE,
  intersect = TRUE,
  geo_graticules = TRUE,
  title = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{...}{Population/geographic region objects of the 'stamp'
class}

\item{pop_facets}{Plot populations in individual panels?}

\item{time_facets}{Plot time snapshots in individual panels?}

\item{intersect}{intersect the population boundaries against landscape
and other geographic boundaries?}

\item{geo_graticules}{Plot axies with lon/lat graticules?}

\item{title}{Plot title}

\item{nrow, ncol}{Number of columns or rows in the facet plot}
}
\description{
If only geographic regions are given, they are colored. If both
them and populations are given, only populations are specified.
}
