% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ancestries}
\alias{plot_ancestries}
\title{Plot simulated ancestry proportions}
\usage{
plot_ancestries(model, output_prefix = file.path(model$path, "output"))
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}

\item{output_prefix}{A shared prefix path to output files (by default, all files
will share a prefix \code{"output"} and are placed in the model directory)}
}
\description{
Plot simulated ancestry proportions
}
