% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot geneflow graph based on given model configuration}
\usage{
plot_graph(model)
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}
}
\description{
Plot geneflow graph based on given model configuration
}
