% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_locations}
\alias{plot_locations}
\title{Plot locations of individuals or nodes in the tree sequence}
\usage{
plot_locations(
  data,
  x = NULL,
  pop = NULL,
  older_than = NULL,
  younger_than = NULL,
  full_scale = TRUE
)
}
\arguments{
\item{x}{Either a character vector of names of sampled individuals, or an or
an integer vector with numeric node IDs}

\item{pop}{Name of the population to plot}

\item{older_than, younger_than}{Time boundaries for the samples}

\item{full_scale}{Plot time gradient on the full scale (spanning the oldest
sampled individual to the present)}

\item{spatial}{Spatial tree sequence data frame (see \code{\link{ts_data}}))}
}
\value{
A ggplot2 object with locations on a map
}
\description{
Plot locations of individuals or nodes in the tree sequence
}
