% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{plot_maps}
\alias{plot_maps}
\title{Plot spatial maps}
\usage{
plot_maps(
  ...,
  time = NULL,
  migrations = FALSE,
  graticules = "original",
  intersect = TRUE,
  show_map = TRUE
)
}
\description{
Plot spatial maps
}
