% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_spatial.R
\name{plot_parents}
\alias{plot_parents}
\title{Plot locations of ancestors of given sample on a map}
\usage{
plot_parents(x, individual)
}
\description{
Plot locations of ancestors of given sample on a map
}
