% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_spatial.R
\name{plot_samples}
\alias{plot_samples}
\title{Plot locations of sampled individuals}
\usage{
plot_samples(data, pop = NULL, older_than = NULL, younger_than = NULL)
}
\arguments{
\item{pop}{Name of the population to plot}

\item{older_than, younger_than}{Time boundaries for the samples}

\item{spatial}{ts data XXXXX}
}
\description{
Plot locations of sampled individuals
}
