% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slendr_map}
\alias{print.slendr_map}
\title{Print a short summary of a \code{slendr_map} object}
\usage{
\method{print}{slendr_map}(x, ...)
}
\arguments{
\item{x}{Object of a class \code{slendr_map}}
}
\description{
All spatial objects in the slendr package are internally
represented by a Simple Features (sf) object. This fact is hidden
in most circumstances this, as the goal of the slendr package is to
provide functionality at a much higher level (population
boundaries, geographic regions, instead of individual polygons and
other "low-level" geometric objects). The logical argument
\code{sf} allows the user to inspect the underlying object
structure. Similarly, by default, \code{print} does not print all
spatial map information to reduce clutter. If many individual
spatial maps for a population are present, a full table of spatial
snapshots can be printed by typing \code{x[]}.
}
