% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slendr_pop}
\alias{print.slendr_pop}
\alias{print.slendr_region}
\alias{print.slendr_map}
\alias{print.slendr_model}
\title{Print a short summary of a \code{slendr} object}
\usage{
\method{print}{slendr_pop}(x, ...)

\method{print}{slendr_region}(x, ...)

\method{print}{slendr_map}(x, ...)

\method{print}{slendr_model}(x, ...)
}
\arguments{
\item{x}{Object of a class \code{slendr} (either \code{slendr_pop},
\code{slendr_map}, \code{slendr_region}, or \code{slendr_table})}

\item{...}{Additional arguments passed to \code{print}}
}
\value{
No return value, used only for printing
}
\description{
All spatial objects in the slendr package are internally represented as
Simple Features (\code{sf}) objects. This fact is hidden in most
circumstances this, as the goal of the slendr package is to provide
functionality at a much higher level (population boundaries, geographic
regions, instead of individual polygons and other "low-level" geometric
objects), without the users having to worry about low-level details involved
in handling spatial geometries. However, the full \code{sf} object
representation can be always printed by calling \code{x[]}.
}
