% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{print.slendr_ts}
\alias{print.slendr_ts}
\title{Print tskit's summary table of the Python tree-sequence object}
\usage{
\method{print}{slendr_ts}(x, ...)
}
\arguments{
\item{x}{Tree object of the class \code{slendr_phylo}}
}
\value{
No return value, simply prints the tskit summary table to the terminal
}
\description{
Print tskit's summary table of the Python tree-sequence object
}
