% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spammr.R
\name{print.spammr}
\alias{print.spammr}
\title{Print a spammr object}
\usage{
\method{print}{spammr}(x, sf = FALSE)
}
\arguments{
\item{x}{Object of a class spammr}

\item{sf}{Print the low-level 'sf' object information?}
}
\description{
Print a spammr object
}
