% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{print.spannr}
\alias{print.spannr}
\title{Print a spannr object}
\usage{
\method{print}{spannr}(x, sf = FALSE, full = FALSE)
}
\arguments{
\item{x}{Object of a class spannr}

\item{sf}{Print the low-level 'sf' object information?}

\item{full}{Print all snapshots? (too many can be bothersome and require
extensive scrolling)}
}
\description{
Print a spannr object
}
