% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{print.stamp}
\alias{print.stamp}
\title{Print a stamp object}
\usage{
\method{print}{stamp}(x, sf = FALSE, full = FALSE)
}
\arguments{
\item{x}{Object of a class stamp}

\item{sf}{Print the low-level 'sf' object information?}

\item{full}{Print all snapshots? (too many can be bothersome and require
extensive scrolling)}
}
\description{
Print a stamp object
}
