% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{raster_center}
\alias{raster_center}
\title{Calculate pixel location of a spatial point (sf object) on a given
raster}
\usage{
raster_center(pop, raster)
}
\arguments{
\item{raster}{Raster object of the class 'stars'}

\item{x}{Point object of the sf type (see \code{get_centroid})}

\item{world}{Underlying world object}
}
\value{
Two-dimensional numeric vector of pixel coordinates
}
\description{
Calculate pixel location of a spatial point (sf object) on a given
raster
}
