% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{rasterize}
\alias{rasterize}
\title{Rasterize the vector form of a population spatial boundary}
\usage{
rasterize(x, resolution)
}
\description{
Rasterize the vector form of a population spatial boundary
}
