% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{read}
\alias{read}
\title{Read a previously serialized spammr model}
\usage{
read(model_dir)
}
\arguments{
\item{model_dir}{Directory with all required configuration files}
}
\description{
Read a previously serialized spammr model
}
