% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_ancestries}
\alias{read_ancestries}
\title{Read ancestry proportion tables generated by the backend SLiM script}
\usage{
read_ancestries(model_dir)
}
\description{
Read ancestry proportion tables generated by the backend SLiM script
}
