% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_example}
\alias{read_example}
\title{Read a provided slendr model}
\usage{
read_example(name = c("introgression", "space"))
}
\arguments{
\item{name}{Name of the built-in slendr model example}
}
\value{
Compiled \code{slendr_model} model object which encapsulates all
information about the specified model (which populations are involved,
when and how much gene flow should occur, what is the spatial resolution
of a map, and what spatial dispersal and mating parameters should be used
in a SLiM simulation, if applicable)
}
\description{
Read a provided slendr model
}
