% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{region}
\alias{region}
\title{Define a geographic region}
\usage{
region(name, map, coords)
}
\arguments{
\item{name}{Name of the geographic region}

\item{map}{Object of the type \code{sf} which defines the map}

\item{coords}{List of vector pairs, defining corners of the range}
}
\value{
Object of the \code{spannr_region} (and \code{sf}) class
}
\description{
Define a geographic region
}
