% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{render_ranges}
\alias{render_ranges}
\title{Take a list of all population regions and intersect them with the
set of underlying world map}
\usage{
render_ranges(ranges, world, boundary = NULL)
}
\description{
Take a list of all population regions and intersect them with the
set of underlying world map
}
