% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{reproject}
\alias{reproject}
\title{Reproject coordinates between coordinate systems}
\usage{
reproject(
  from,
  to,
  x = NULL,
  y = NULL,
  coords = NULL,
  model = NULL,
  add = FALSE,
  input_prefix = "",
  output_prefix = "new"
)
}
\arguments{
\item{from, to}{Either a CRS code accepted by GDAL, a valid integer
EPSG value, an object of class \code{crs}, the value "raster"
(converting from/to pixel coordinates), or "world" (converting
from/to whatever CRS is set for the underlying map)}

\item{x, y}{Coordinates in two dimensions (if missing, coordinates
are expected to be in the \code{data.frame} specified in the
\code{coords} parameter as columns "x" and "y")}

\item{coords}{data.frame-like object with coordinates in columns
"x" and "y"}

\item{model}{Object of the class \code{slendr_model}}

\item{add}{Add column coordinates to the input data.frame
\code{coords} (coordinates otherwise returned as a separate
object)?}
}
\value{
Data.frame with converted two-dimensional coordinates
}
\description{
Converts between coordinates on a compiled raster map (i.e. pixel
units) and different Geographic Coordinate Systems (CRS).
}
