% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{sampling}
\alias{sampling}
\title{Define sampling events at specified times for a given set of populations}
\usage{
sampling(times, ..., strict = FALSE)
}
\arguments{
\item{times}{Integer vector of times (in model time units) at which to
schedule remembering of individuals in the tree-sequence}

\item{...}{Lists of two elements (\code{slendr_pop} population object-<number
of individuals to sample), representing from which populations should how
many individuals be remembered at times given by \code{times}}
}
\description{
Define sampling events at specified times for a given set of populations
}
