% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{save_png}
\alias{save_png}
\title{Save the rasterized stars object to a PNG file}
\usage{
save_png(raster, path)
}
\arguments{
\item{raster}{Object of a stars type}

\item{path}{Full path of an output PNG file}
}
\description{
Save the rasterized stars object to a PNG file
}
