% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{script}
\alias{script}
\title{Substitute variables in a template SLiM script}
\usage{
script(path, output = NULL, ...)
}
\arguments{
\item{path}{Path to a template SLiM script}

\item{output}{Where to save the substituted SLiM script (temporary location
by default)}

\item{...}{Variable values to be substituted}
}
\value{
Name of the substituted SLiM script
}
\description{
Variables in the template script must conform to the Mustache
specification, i.e. specified as {{variable_name}}. For more
information see the Mustache specification at:
\url{http://mustache.github.io/}
}
