% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ranges.R, R/spannr.R
\name{set_class}
\alias{set_class}
\title{Set spannr classes (or fix their priorities if already present)}
\usage{
set_class(x, type)

set_class(x, type)
}
\description{
Set spannr classes (or fix their priorities if already present)

Set spannr classes (or fix their priorities if already present)
}
