% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slendr-package.R
\docType{package}
\name{slendr}
\alias{slendr}
\alias{slendr-package}
\title{Population Genetics Simulations Spanning Time And Space}
\description{
The \emph{slendr} package provides a framework for simulating
spatially-explicit genomic data. It leverages real Earth cartographic
information (or, alternatively, any kind of abstract spatial landscape) to
programmatically and visually encode spatial population boundaries and
temporal dynamics (i.e., population displacements and expansions) across a
geographic landscape, and specify population divergences and gene flow
events. Spatial demographic models are then compiled into a custom-built SLiM
script which can be extended by the user to perform arbitrary simulations.
}
\details{
You can find installation instructions, reference manual, and tutorials at
\url{https://bodkan.net/slendr}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bodkan/slendr}
  \item Report bugs at \url{https://github.com/bodkan/slendr/issues}
}

}
\author{
\strong{Maintainer}: Martin Petr \email{contact@bodkan.net} (\href{https://orcid.org/0000-0003-4879-8421}{ORCID})

}
\keyword{internal}
