% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{substitute}
\alias{substitute}
\title{Substitute values of parameters in a given slendr/SLiM extension}
\usage{
substitute(template, ...)
}
\arguments{
\item{template}{Either a path to an extension script file, or a string
containing the entire SLiM extension code}

\item{...}{Named function arguments interpreted as key=value pairs to be
used in argument substitution}
}
\value{
Path to a file with a saved extension script containing all
substituted values
}
\description{
Substitute values of parameters in a given slendr/SLiM extension
}
