% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.slendr_nodes}
\alias{summary.slendr_nodes}
\title{Summarise the contents of a \code{ts_nodes} result}
\usage{
\method{summary}{slendr_nodes}(object, ...)
}
\arguments{
\item{object}{Data frame produced by the function \code{ts_nodes}}

\item{...}{Additional formal arguments to the \code{summary} method (unused here)}
}
\value{
Used for its output to the terminal
}
\description{
Summarise the contents of a \code{ts_nodes} result
}
