% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_direction}
\alias{time_direction}
\title{Get direction of time implied by the history of the population}
\usage{
time_direction(x)
}
\arguments{
\item{x}{Object of the class \code{slendr_pop} or \code{slendr_model}}
}
\value{
Either "forward", "backward", or "unknown"
}
\description{
Get direction of time implied by the history of the population
}
