% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_spatial.R
\name{ts_ancestors}
\alias{ts_ancestors}
\title{Extract the ancestral relationships for a set of nodes}
\usage{
ts_ancestors(x, data, ts)
}
\arguments{
\item{x}{Either a string representing an individual name, or an integer
number specifying a node in a tree sequence}

\item{data}{Object of the class \code{slendr_spatial} carrying the
spatio-temporal position of each individual node in the tree sequence}

\item{ts}{pyslim.SlimTreeSequence object of the class \code{slendr_ts}
obtained by \code{link{ts_load}}, \code{link{ts_recapitate}},
\code{link{ts_simplify}}, or \code{link{ts_mutate}}}
}
\description{
Extract the ancestral relationships for a set of nodes
}
