% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_branches}
\alias{ts_branches}
\title{Extract table of spatio-temporal node annotation from a given tree}
\usage{
ts_branches(tree)
}
\arguments{
\item{tree}{Tree object generated by \code{ts_phylo}}
}
\value{
Data frame of the \code{sf} type containing the times of nodes and
start-end coordinates of branches across space
}
\description{
Extract table of spatio-temporal node annotation from a given tree
}
