% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_data}
\alias{ts_data}
\title{Extract combined annotated table of individuals and nodes}
\usage{
ts_data(ts, remembered = NULL, retained = NULL, alive = NULL)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{remembered, retained, alive}{Only extract information about nodes and
times of individuals with the specific flag}
}
\value{
Data frame with processed information from the tree sequence object.
If the model which generated this data was spatial, result will be returned
as a spatial object of the class \code{sf}. See description for more
information.
}
\description{
This function combines information from the table of individuals and table of
nodes into a single data frame which can be used in downstream analyses.
}
\details{
The source of data (tables of individuals and nodes recorded in the tree
sequence generated by SLiM) are combined into a single data frame. If the
model which generated the data was spatial, coordinates of nodes (which are
pixel-based by default because SLiM spatial simulations occur on a raster),
the coordinates are automatically converted to an explicit spatial object of
the \code{sf} class unless \code{spatial = FALSE}. See
\url{https://r-spatial.github.io/sf/} for an extensive introduction to the sf
package and the ways in which spatial data can be processed, analysed, and
visualised.
}
\seealso{
\code{\link{ts_individuals}} \code{\link{ts_nodes}}
\code{\link{ts_edges}} for accessing raw tree sequence tables without
added metadata annotation. See also \code{\link{ts_ancestors}} to learn
how to extract information about relationship beteween nodes in the tree
sequence, and how to analysed data about distances between nodes in the
spatial context.
}
