% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_descendants}
\alias{ts_descendants}
\title{Extract all descendants of a given tree-sequence node}
\usage{
ts_descendants(ts, x, verbose = FALSE)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{x}{An integer node ID of the ancestral node}

\item{verbose}{Report on the progress of ancestry path generation?}
}
\description{
Extract all descendants of a given tree-sequence node
}
