% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_diversity}
\alias{ts_diversity}
\title{Calculate diversity in given sets of individuals}
\usage{
ts_diversity(
  ts,
  sample_sets,
  mode = c("site", "branch", "node"),
  windows = NULL,
  span_normalise = TRUE
)
}
\arguments{
\item{ts}{\code{pyslim.SlimTreeSequence} object}

\item{sample_sets}{A list (optionally a named list) of character vectors with
individual names (one vector per set)}

\item{mode}{The mode for the calculation ("sites" or "branch")}

\item{windows}{Coordinates of breakpoints between windows. The first
coordinate (0) and the last coordinate (equal to \code{ts$sequence_length})
are added automatically)}

\item{span_normalise}{Divide the result by the span of the window? Default
TRUE, see the tskit documentation for more detail.}
}
\value{
For each set of individuals either a single diversity value or a
vector of diversity values (one for each window)
}
\description{
Calculate diversity in given sets of individuals
}
