% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_draw}
\alias{ts_draw}
\title{Plot a graphical representation of a tree sequence or a single tree}
\usage{
ts_draw(x, width = 1500, height = 500, individuals = FALSE, ts = NULL, ...)
}
\arguments{
\item{x}{A tree sequence loaded by \code{ts_load} or a single tree extracted
from a tree sequence by \code{ts_tree}}

\item{width, height}{Pixel dimensions of the rendered bitmap}

\item{...}{Keyword arguments to the tskit \code{draw_svg} function.}
}
\description{
This function obtains an SVG representation by calling the \code{draw_svg}
method of tskit and renders it as a bitmap imagine in R. All of the many
optional keyword arguments can be provided and will be automatically passed
to \code{draw_svg} behind the scenes.
}
