% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_eigenstrat}
\alias{ts_eigenstrat}
\title{Save genotypes from the tree-sequence in an EIGENSTRAT format}
\usage{
ts_eigenstrat(ts, prefix)
}
\arguments{
\item{ts}{SlimTreeSequence object}

\item{prefix}{EIGENSTRAT trio prefix}
}
\value{
Object of the class EIGENSTRAT created by the admixr package
}
\description{
Save genotypes from the tree-sequence in an EIGENSTRAT format
}
