% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_extend}
\alias{ts_extend}
\title{Extend the span of sequence covered by ancestral nodes (EXPERIMENTAL)}
\usage{
ts_extend(ts, iterations = 10)
}
\arguments{
\item{ts}{Tree sequence object}

\item{iterations}{Maximum number of iterations over the tree sequence
(default is 10)}
}
\value{
Tree-sequence object of the class \code{slendr_ts}
}
\description{
Perform the "extend haplotypes" algorithm to extend the span of sequence
covered by ancestral nodes as implemented by the tskit method
\code{TreeSequence.extend_haplotypes}.
}
\details{
For more details about this functionality, please read the following
resources:
\enumerate{
\item Paper by Fritz et al. 2025 (\doi{10.1093/genetics/iyaf198})
\item Official tskit documentation at
\url{https://tskit.dev/tskit/docs/stable/python-api.html#tskit.TreeSequence.extend_haplotypes}
}

Note that this function has been currently tested exclusively on tree
sequences simulated by slendr's simulation engine implemented by the function
\code{msprime()} executed with \code{coalescent_only = FALSE}.
}
\examples{
\dontshow{check_dependencies(python = TRUE, quit = TRUE) # dependencies must be present
}
init_env()

# load an example model with an already simulated tree sequence
slendr_ts <- system.file("extdata/models/introgression_msprime.trees", package = "slendr")
model <- read_model(path = system.file("extdata/models/introgression", package = "slendr"))

ts <- ts_read(slendr_ts, model)
ts_ext <- ts_extend(ts, iterations = 10)

ts_ext
}
