% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_f2}
\alias{ts_f2}
\alias{ts_f3}
\alias{ts_f4}
\alias{ts_f4ratio}
\title{Calculate the f2, f3, f4, and f4-ratio statistics}
\usage{
ts_f2(ts, a, b, ...)

ts_f3(ts, a, b, c, ...)

ts_f4(ts, a, b, c, d, ...)

ts_f4ratio(ts, x, a, b, c, d)
}
\arguments{
\item{ts}{SlimTreeSequence object}

\item{x, a, b, c, d}{Character vectors of individual names}
}
\value{
Numeric estimate of ancestry proportion
}
\description{
Calculate the f2, f3, f4, and f4-ratio statistics
}
