% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_ibd}
\alias{ts_ibd}
\title{Extract Identity-by-Descent (IBD) segments (EXPERIMENTAL)}
\usage{
ts_ibd(ts, within = NULL, between = NULL, maximum_time = NULL, sf = TRUE)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{within}{A character vector with individual names or an integer vector with
node IDs indicating a set of nodes within which to look for IBD segments.}

\item{between}{A list of lists of character vectors with individual names or
integer vectors with node IDs, indicating a set of nodes between which to
look for shared IBD segments.}

\item{maximum_time}{Oldest MRCA of a node to be considered as an IBD ancestor
to return that IBD segment in results. This is useful for reducing the total
amount of IBD returned.}

\item{sf}{If IBD segments in a spatial tree sequence are being analyzed, should
the returned table be a spatial sf object? Default is \code{TRUE}.}
}
\value{
A data frame with IBD coordinates of each IBD segment shared by a
pair of nodes
}
\description{
This function extracts IBD segments between pairs of individuals or nodes
from a given tree sequence.
}
\details{
This function is considered highly experimental. For full control over IBD segment
detection in tree-sequence data, users can (and perhaps, for the time being,
should) rely on the tskit method \code{ibd_segments}
(see \url{https://tskit.dev/tskit/docs/stable/python-api.html#tskit.TreeSequence.ibd_segments})
rather than on this R function.

This is particularly because of the fact that, similarly to many other
tskit-based functions provided by slendr, \code{ts_ibd()} has a very
opinionated interface and makes strong assumptions about the kind of IBD
segments the user is interested in extracting. Briefly, \code{ts_ibd()}
only extracts IBD segments in a "squashed" form, and thus does not provide
means to filter out IBD segments based on their lengths as one might do
in a tskit's own assumption of what is considered to be a unique IBD segment.
See \url{https://github.com/tskit-dev/tskit/issues/2459} for more context.
Furthermore, in order to return IBD segments as close to their traditional
definition (which is slightly different from tskit's own definition), this
function implicitly assumes that the given tree sequence has been processed
by the "extend haplotypes" algorithm as described by Fritze et al. 2025
(\doi{10.1093/genetics/iyaf198}) and available under the slendr function
\code{ts_extend()}.
}
\examples{
\dontshow{check_dependencies(python = TRUE, quit = TRUE) # dependencies must be present
}
init_env()

# load an example model with an already simulated tree sequence
slendr_ts <- system.file("extdata/models/introgression_slim.trees", package = "slendr")
model <- read_model(path = system.file("extdata/models/introgression", package = "slendr"))

# load the tree-sequence object from disk
ts <- ts_read(slendr_ts, model)

# find IBD segments between specified Neanderthals and Europeans
ts_ibd(
  ts,
  coordinates = TRUE,
  between = list(c("NEA_1", "NEA_2"), c("EUR_1", "EUR_2")),
  minimum_length = 40000
)
}
\seealso{
\code{\link{ts_extend}} for information about the "extend haplotypes"
procedure
}
