% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_individual}
\alias{ts_individual}
\title{Extract a single individual from a tree-sequence}
\usage{
ts_individual(ts, id = NULL, pedigree_id = NULL, name = NULL)
}
\arguments{
\item{ts}{SlimTreeSequence object}

\item{id}{Numeric index of an individual}

\item{pedigree_id}{Pedigree number of an individual}

\item{name}{String name of the individual (all names can be accessed by \code{ts_individuals})}
}
\value{
Object of the type \code{tskit.trees.Individual}
}
\description{
Extract a single individual from a tree-sequence
}
