% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_individuals}
\alias{ts_individuals}
\title{Extract table of individuals from pyslim's tree-sequence object}
\usage{
ts_individuals(ts)
}
\arguments{
\item{ts}{SlimTreeSequence object}

\item{model}{\code{slendr_model} object of a model that generated the data}
}
\value{
Data frame with individual information from the input tree-sequence
object
}
\description{
Extract table of individuals from pyslim's tree-sequence object
}
