% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_mutate}
\alias{ts_mutate}
\title{Add mutations to the given tree-sequence}
\usage{
ts_mutate(ts, mutation_rate)
}
\arguments{
\item{ts}{Object of the type SlimTreeSequence}

\item{mutation_rate}{Mutation rate used by msprime to simulate mutations}
}
\value{
SlimTreeSequence object
}
\description{
Add mutations to the given tree-sequence
}
