% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_nodes}
\alias{ts_nodes}
\title{Extract the nodes belonging to the given individuals}
\usage{
ts_nodes(ts, ind)
}
\arguments{
\item{ts}{SlimTreeSequence object}

\item{ind}{A character vector of individual names}
}
\value{
An integer vector with node numbers
}
\description{
Extract the nodes belonging to the given individuals
}
