% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_recapitate}
\alias{ts_recapitate}
\title{Recapitate the tree-sequence}
\usage{
ts_recapitate(ts, recombination_rate, Ne)
}
\arguments{
\item{ts}{SlimTreeSequence object loaded by \code{ts_load}}

\item{recombination_rate}{A constant value of the recombination rate}

\item{Ne}{Effective population size during the recapitation simulation}
}
\value{
@return SlimTreeSequence object accessed via the reticulate package
}
\description{
Recapitate the tree-sequence
}
