% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_save}
\alias{ts_save}
\title{Write a tree sequence to a file}
\usage{
ts_save(ts, file)
}
\arguments{
\item{file}{A path to the tree-sequence file (either originating from a
slendr model or a standard non-slendr tree sequence).}

\item{model}{Optional \code{slendr_model} object which produced the
tree-sequence \code{file}. Used for adding various annotation data and
metadata to the standard tskit tree-sequence object.}
}
\description{
Deprecated function. Please use \code{ts_write} instead.
}
