% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_segregating}
\alias{ts_segregating}
\title{Calculate the density of segregating sites for the given sets of individuals}
\usage{
ts_segregating(
  ts,
  sample_sets,
  mode = c("site", "branch", "node"),
  windows = NULL,
  span_normalise = TRUE
)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{sample_sets}{A list (optionally a named list) of character vectors with
individual names (one vector per set). If a simple vector is provided, it
will be interpreted as \code{as.list(sample_sets)}, meaning that a given
statistic will be calculated for each individual separately.}

\item{mode}{The mode for the calculation ("sites" or "branch")}

\item{windows}{Coordinates of breakpoints between windows. The first
coordinate (0) and the last coordinate (equal to \code{ts$sequence_length})
are added automatically)}

\item{span_normalise}{Divide the result by the span of the window? Default
TRUE, see the tskit documentation for more detail.}
}
\value{
For each set of individuals either a single diversity value or a
vector of diversity values (one for each window)
}
\description{
Calculate the density of segregating sites for the given sets of individuals
}
