% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_spatial}
\alias{ts_spatial}
\title{Convert spatiotemporal tree-sequence data to a spatial format}
\usage{
ts_spatial(ts, crs = NULL)
}
\arguments{
\item{ts}{SlimTreeSequence object}

\item{crs}{For spatial models, which coordinate reference system to reproject
individual locations to? If \code{NULL} (default), coordinates are
reprojected to the coordinate reference system (CRS) of the entire spatial
model. Alternatively, a CRS code accepted by GDAL, a valid integer EPSG
value, an object of class \code{crs}, can be specified to change the
reprojection to a different CRS.}
}
\value{
Spatial object of the class \code{slendr_ts} and \code{sf}
}
\description{
Convert spatiotemporal tree-sequence data to a spatial format
}
