% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_tajima}
\alias{ts_tajima}
\title{Calculate Tajima's D for a group of individuals}
\usage{
ts_tajima(ts, sample, mode = c("site", "branch"), windows = NULL)
}
\arguments{
\item{mode}{The mode for the calculation ("sites" or "branch")}

\item{windows}{Breakpoint coordinates between windows (first coordinate 0 and
last coordinate equal to \code{ts$sequence_length} are added automatically)}

\item{tsSlimTreeSequence}{object}

\item{set}{Character vector of individual names}
}
\value{
Either a single Fst value or a vector of Fst values (one for each window)
}
\description{
Calculate Tajima's D for a group of individuals
}
