% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{world}
\alias{world}
\title{Define a world map for all spatial operations}
\usage{
world(xrange, yrange, landscape = "naturalearth", crs = NULL, ne_dir = NULL)
}
\arguments{
\item{xrange}{Numeric vector with minimum and maximum longitude}

\item{yrange}{Numeric vector with minimum and maximum latitude}

\item{landscape}{Either "blank" (for blank abstract geography),
"naturalearth" (for real Earth geography) or an object of the
class \code{sf} defining abstract geographic features of the
world}

\item{crs}{EPSG code of a coordinate reference system to use for
spatial operations (no CRS assumed by default, implying an
abstract landscape not tied to any real-world geographic region)}

\item{ne_dir}{Path to the directory where Natural Earth data was
manually downloaded and unzipped (used only when \code{landscape}
is "naturalearth")}
}
\value{
Object of the class \code{spannr_map}
}
\description{
Download Natural Earth land area map data, zoom on a given window
determined by longitude and latitude coordinates and transform to a
specified projection. Alternatively, load a previously downloaded
and unzipped data such as the "Land polygons" data here:
\url{https://www.naturalearthdata.com/downloads/110m-physical-vectors/}
}
