% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{world_map}
\alias{world_map}
\title{Define a world map for all spatial operations}
\usage{
world_map(lon_range, lat_range, crs = "EPSG:4326")
}
\arguments{
\item{lon_range}{Numeric vector with minimum and maximum longitude}

\item{lat_range}{Numeric vector with minimum and maximum latitude}

\item{crs}{Coordinate Reference System to use for all spatial
operations (default is WGS-84 or EPSG:4326 CRS)}
}
\description{
Download Natural Earth land area map data, zoom on a given window
determined by longitude and latitude coordinates and transform to a
specified projection
}
