% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{write}
\alias{write}
\title{Write a compiled spannr model to disk}
\usage{
write(model, dir, overwrite)
}
\arguments{
\item{model}{Compiled \code{spannr_model} model object}

\item{dir}{Output directory for the model configuration files which
will be loaded by the backend SLiM script}

\item{overwrite}{Overwrite the contents of the output directory in
case it already exists?}
}
\description{
Write a compiled spannr model to disk
}
