% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{write_maps}
\alias{write_maps}
\title{Save the rasterized spatial maps as PNG files into a target
directory}
\usage{
write_maps(maps_table, output_dir)
}
\description{
Save the rasterized spatial maps as PNG files into a target
directory
}
