% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{write_model}
\alias{write_model}
\title{Write model specification tables which will be loaded into SLiM}
\usage{
write_model(model_dir, splits_table, admix_table, maps_table, gen_time)
}
\description{
Write model specification tables which will be loaded into SLiM
}
