% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{admixture}
\alias{admixture}
\title{Define an admixture event}
\usage{
admixture(from, to, rate, start, end, overlap = 0.1)
}
\arguments{
\item{from, to}{Population range objects of the class \code{spammr_pop}}

\item{rate}{Scalar value in the range (0, 1] specifying the proportion of
migration over given time period}

\item{start, end}{Start and end of the admixture event}

\item{overlap}{Proportion of the spatial range of the migrating population
\code{from} required to overlap with the range of the \code{to} population.
Value \code{FALSE} disables the overlap requirement, allowing admixture
between non-overlapping populations.}
}
\value{
Object of the class data.frame
}
\description{
Define an admixture event
}
