% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{admixture}
\alias{admixture}
\title{Define an admixture event}
\usage{
admixture(from, to, rate, start, end)
}
\arguments{
\item{from, to}{Population range objects of the class \code{spammr_pop}}

\item{rate}{Scalar value in the range (0, 1] specifying the proportion of
migration over given time period}

\item{start, end}{Start and end of the admixture event}
}
\value{
Object of the class data.frame
}
\description{
Define an admixture event
}
