% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ancestries}
\alias{ancestries}
\title{Plot simulated ancestry proportions}
\usage{
ancestries(output_dir, gen_time = NULL)
}
\arguments{
\item{output_dir}{Directory with simulation output files}
}
\description{
Plot simulated ancestry proportions
}
