% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{animate}
\alias{animate}
\title{Animate the simulated population dynamics}
\usage{
animate(model, nframes = 200, gif = NULL)
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}

\item{nframes}{Number of frames of the animation}

\item{gif}{Path to an output GIF (animation object returned by default)
#' @param nframes Number of frames of the animation}

\item{locations}{Table with individual locations}
}
\value{
If \code{gif = NULL}, return gganimate animation object. Otherwise a GIF
file is saved and no value is returned.
}
\description{
Animate the simulated population dynamics
}
