% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{animate}
\alias{animate}
\title{Animate the simulated population dynamics}
\usage{
animate(model, nframes, gif = NULL)
}
\arguments{
\item{model}{Compiled model object}

\item{nframes}{Number of frames of the animation}

\item{gif}{Path to an output GIF (animation object returned by default)
#' @param nframes Number of frames of the animation}
}
\value{
If \code{gif = NULL}, return gganimate animation object. Otherwise a GIF
file is saved and no value is returned.
}
\description{
Animate the simulated population dynamics
}
