% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{change}
\alias{change}
\title{Add a new time snapshot for an already defined population}
\usage{
change(
  pop,
  time,
  Ne = NULL,
  center = NULL,
  radius = NULL,
  coords = NULL,
  region = NULL
)
}
\arguments{
\item{pop}{Population object of the \code{spammr} class}

\item{time}{Time of the current snapshot that is being defined}

\item{Ne}{Effective population size (stays the same by default)}

\item{center}{Vector of two elements defining a center of a circular range}

\item{radius}{Scalar defining a radius of a range in kilometers}

\item{coords}{List of vector pairs, defining corners of the range}

\item{region}{Geographic region of the class \code{spammr_region}}
}
\value{
Object of the \code{spammr_pop} (and \code{sf}) class
}
\description{
This function allows a more manual control of spatial map changes in addition
to the \code{expand} and \code{move} functions
}
