% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{change}
\alias{change}
\title{Update the population map or one of its parameters}
\usage{
change(pop, time, N = NULL, center = NULL, radius = NULL, polygon = NULL)
}
\arguments{
\item{pop}{Object of the class \code{spannr_pop}}

\item{time}{Time of the change}

\item{N}{Number of individuals}

\item{center}{Two-dimensional vector specifying the center of the
circular range}

\item{radius}{Radius of the circular range}

\item{polygon}{List of vector pairs, defining corners of the
polygon range (see also the \code{region} argument) or a
geographic region of the class \code{spannr_region} from which
the polygon coordinates will be extracted}
}
\value{
Object of the class \code{spannr_pop}
}
\description{
This function allows a more manual control of spatial map changes
in addition to the \code{expand} and \code{move} functions
}
