% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check that all dependencies are available for slendr examples}
\usage{
check_dependencies(python = FALSE, slim = FALSE)
}
\arguments{
\item{python}{Is the slendr Python environment required?}

\item{slim}{Is SLiM required?}
}
\value{
No return value. Called only to result in an error message if a
particular software dependency is missing for an example to run.
}
\description{
Check that all dependencies are available for slendr examples
}
