% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{check_env}
\alias{check_env}
\title{Check that the active Python environment is correctly setup for slendr}
\usage{
check_env()
}
\description{
This function inspects the Python environment which has been activated by the
reticulate package and prints the versions of all slendr Python dependencies
to the console.
}
