% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{check_env}
\alias{check_env}
\title{Check that the currently active Python environment is setup correctly}
\usage{
check_env()
}
\description{
Extracts the components of the Python environment that would be currently
used by reticulate and prints versions of all Python dependencies.
}
