% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{clear_env}
\alias{clear_env}
\title{Remove the automatically created slendr Python environment}
\usage{
clear_env(force = FALSE)
}
\arguments{
\item{force}{Ask before deleting the environment?}
}
\description{
Remove the automatically created slendr Python environment
}
