% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{compile}
\alias{compile}
\title{Compile the spatial maps and split/admixture tables}
\usage{
compile(..., dry_run = T)
}
\arguments{
\item{...}{Population object of the 'spammr_pop' class}

\item{dry_run}{Return the compiled model components without saving them to
disk?}
}
\description{
Compile the spatial maps and split/admixture tables
}
