% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{compile}
\alias{compile}
\title{Deprecated function}
\usage{
compile(
  populations,
  generation_time,
  path = NULL,
  resolution = NULL,
  competition = NULL,
  mating = NULL,
  dispersal = NULL,
  geneflow = list(),
  overwrite = FALSE,
  sim_length = NULL,
  direction = NULL,
  slim_script = system.file("scripts", "script.slim", package = "slendr"),
  description = "",
  dir = NULL
)
}
\description{
Deprecated function
}
