% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{compile}
\alias{compile}
\title{Compile the spatial maps and split/admixture tables}
\usage{
compile(..., outdir, overwrite = FALSE)
}
\arguments{
\item{...}{Population object of the 'spammr_pop' class}

\item{outdir}{Directory to which to save the entire model}

\item{overwrite}{Overwrite the contents of the output directory (in case it
already exists)?}
}
\description{
Compile the spatial maps and split/admixture tables
}
