% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{compile}
\alias{compile}
\title{Compile the spatial demographic model}
\usage{
compile(
  populations,
  dir,
  generation_time,
  resolution,
  admixtures = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{populations}{Object(s) of the \code{spannr_pop} class
(multiple objects need to be specified in a \code{list})}

\item{dir}{Output directory for the model configuration files which
will be loaded by the backend SLiM script}

\item{generation_time}{Generation time (in model time units)}

\item{resolution}{How many distance units per pixel?}

\item{admixtures}{Admixture events generated by the
\code{admixture} function (either a list of data.frame objects in
the format defined by the \code{admixture} function, or a single
data.frame)}

\item{overwrite}{Overwrite the contents of the output directory in
case it already exists?}
}
\value{
Compiled \code{spannr_model} model object
}
\description{
First, compiles the vectorized population spatial maps into a
series of binary raster PNG files, which is the format that SLiM
understands and uses it to define population boundaries. Then
extracts the demographic model defined by the user (i.e. population
divergences and gene flow events) into a series of tables which are
later used by the built-in SLiM script to program the timing of
simulation events.
}
