% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{compile}
\alias{compile}
\title{Compile the spatial demographic model}
\usage{
compile(
  populations,
  dir,
  generation_time,
  resolution = NULL,
  competition_dist = NULL,
  mate_dist = NULL,
  dispersal_dist = NULL,
  geneflow = list(),
  overwrite = FALSE,
  sim_length = NULL,
  direction = NULL,
  script_path = system.file("extdata", "script.slim", package = "slendr"),
  description = ""
)
}
\arguments{
\item{populations}{Object(s) of the \code{slendr_pop} class
(multiple objects need to be specified in a \code{list})}

\item{dir}{Output directory for the model configuration files which
will be loaded by the backend SLiM script}

\item{generation_time}{Generation time (in model time units)}

\item{resolution}{How many distance units per pixel?}

\item{competition_dist, mate_dist}{Maximum spatial competition and
mating choice distance}

\item{dispersal_dist}{Standard deviation of the normal distribution
of the parent-offspring distance}

\item{geneflow}{Geneflow events generated by the \code{geneflow}
function (either a list of data.frame objects in the format
defined by the \code{geneflow} function, or a single data.frame)}

\item{overwrite}{Overwrite the contents of the output directory in
case it already exists?}

\item{sim_length}{Total length of the simulation (required for
forward time models, optional for models specified in backward
time units which by default run to "the present time")}

\item{direction}{Intended direction of time. Under normal
circumstances this parameter is inferred from the model and does
not need to be set manually.}

\item{script_path}{Path to a SLiM script to be used for executing
the model (by default, a bundled backend script will be used)}

\item{description}{Optional short description of the model}
}
\value{
Compiled \code{slendr_model} model object
}
\description{
First, compiles the vectorized population spatial maps into a series of
binary raster PNG files, which is the format that SLiM understands and uses
it to define population boundaries. Then extracts the demographic model
defined by the user (i.e. population divergences and gene flow events) into a
series of tables which are later used by the built-in SLiM script to program
the timing of simulation events.
}
