% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{compile}
\alias{compile}
\title{Compile the spatial maps and split/admixture tables}
\usage{
compile(
  populations,
  model_dir,
  gen_time,
  resolution,
  admixtures = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{populations}{Population map objects of the \code{stamp_pop} class (a
list of such object or a single element)}

\item{model_dir}{Directory to which to save the entire model}

\item{gen_time}{Generation time (in model's time units, i.e. years)}

\item{resolution}{How many kilometers per pixel?}

\item{admixtures}{Admixture events generated by the \code{admixture} function
(either a list of data.frame objects - one per admixture event, or a single
data.frame for one admixture event)}

\item{overwrite}{Overwrite the contents of the output directory (in case it
already exists)?}
}
\description{
Compile the spatial maps and split/admixture tables
}
