% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{convert}
\alias{convert}
\title{Convert between coordinate system (raster-coordinates or CRS)}
\usage{
convert(from, to, x = NULL, y = NULL, coords = NULL, model = NULL, add = FALSE)
}
\arguments{
\item{from, to}{Either a string accepted by GDAL, a valid integer
EPSG value, an object of class crs, or the value "raster"}

\item{x, y}{Coordinates in two dimensions (if missing, coordinates
expected in the \code{coords} parameter as columns "x" and "y")}

\item{coords}{data.frame-like object with coordinates in columns "x"
and "y"}

\item{model}{object of the class \code{spammr_model}}

\item{add}{Add column coordinates to the input data.frame \code{coords}?}
}
\value{
Data.frame with converted two-dimensional coordinates
}
\description{
Convert between coordinate system (raster-coordinates or CRS)
}
