% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{create_polygon}
\alias{create_polygon}
\title{Create a simple geometry polygon object from the list of
coordinates}
\usage{
create_polygon(coords, world)

create_polygon(coords, world)
}
\description{
Create a simple geometry polygon object from the list of
coordinates

Create a simple geometry polygon object from the list of
coordinates
}
