% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dimension}
\alias{dimension}
\title{Return the dimensions of the world map}
\usage{
dimension(map)
}
\arguments{
\item{map}{Object of the type \code{slendr_map}}
}
\value{
If the coordinate reference system was specified, the function
returns a two-dimensional vector of the world dimensions in the projected
units. Otherwise the dimensions of the two-dimensional abstract plane are
returned.
}
\description{
Return the dimensions of the world map
}
