% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{dispersal}
\alias{dispersal}
\title{Deprecated function}
\usage{
dispersal(
  pop,
  time,
  competition = NA,
  mating = NA,
  dispersal = NA,
  dispersal_fun = NULL
)
}
\description{
Deprecated function
}
