% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dispersal}
\alias{dispersal}
\title{Change dispersal parameters}
\usage{
dispersal(
  pop,
  time,
  competition_dist = NA,
  mate_dist = NA,
  dispersal_dist = NA,
  dispersal_fun = NULL
)
}
\arguments{
\item{pop}{Object of the class \code{slendr_pop}}

\item{time}{Time of the population size change}

\item{competition_dist, mate_dist}{Maximum spatial competition and mating
choice distance}

\item{dispersal_dist}{Standard deviation of the normal distribution of the
distance that offspring disperses from its parent}

\item{dispersal_fun}{Distribution function governing the dispersal of
offspring. One of "normal", "uniform", "cauchy", or "exponential".}
}
\description{
Changes either the competition interactive distance, mating choice distance,
or the dispersal of offspring from its parent
}
