% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{expand}
\alias{expand}
\title{Expand population radius by a given factor in a given time}
\usage{
expand(pop, by, duration, snapshots, region = NULL)
}
\description{
Expand population radius by a given factor in a given time
}
