% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{expand}
\alias{expand}
\title{Expand population radius by a given factor in a given time}
\usage{
expand(region, by, duration, snapshots)
}
\description{
Expand population radius by a given factor in a given time
}
