% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{expand}
\alias{expand}
\title{Expand the population range}
\usage{
expand(pop, by, end, snapshots, start = NULL, polygon = NULL)
}
\arguments{
\item{pop}{Object of the class \code{spannr_pop}}

\item{by}{How many units of distance to expand by?}

\item{snapshots}{Number of time slices to split the expansion into}

\item{start, end}{When does the expansion start/end?}

\item{polygon}{Geographic region to restrict the expansion to}
}
\value{
Object of the class \code{spannr_pop}
}
\description{
Expands the spatial population range by a specified distance in a
given time-window
}
