% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{expand}
\alias{expand}
\title{Expand population radius by a given factor in a given time}
\usage{
expand(pop, by, end, snapshots, start = NULL, region = NULL)
}
\arguments{
\item{pop}{Spatial population object of \code{spannr_pop} class}

\item{by}{How many kilometers to expand by?}

\item{snapshots}{Number of time slices to split the movement into}

\item{start, end}{When does the spatial population expansion start/end?}

\item{region}{Geographic region to restrict the expansion to}
}
\value{
Object of the \code{spannr_pop} (and \code{sf}) class
}
\description{
Expand population radius by a given factor in a given time
}
