% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{expand}
\alias{expand}
\title{Deprecated function}
\usage{
expand(
  pop,
  by,
  end,
  start,
  overlap = 0.8,
  snapshots = NULL,
  polygon = NULL,
  lock = FALSE,
  verbose = TRUE
)
}
\description{
Deprecated function
}
