% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{explore}
\alias{explore}
\title{Open an interactive browser of the spatial model}
\usage{
explore(model, step = model$generation_time)
}
\arguments{
\item{model}{Compiled \code{spannr_model} model object}
}
\description{
Open an interactive browser of the spatial model
}
