% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{geneflow}
\alias{geneflow}
\title{Define a geneflow event}
\usage{
geneflow(from, to, rate, start, end, overlap = TRUE)
}
\arguments{
\item{from, to}{Objects of the class \code{spannr_pop}}

\item{rate}{Scalar value in the range (0, 1] specifying the
proportion of migration over given time period}

\item{start, end}{Start and end of the geneflow event}

\item{overlap}{Require spatial overlap between admixing
populations?  (default \code{TRUE})}
}
\value{
Object of the class data.frame
}
\description{
Define a geneflow event
}
