% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_intermediate_edges}
\alias{get_intermediate_edges}
\title{Create a table of 'intermediate' edges for graph visualization}
\usage{
get_intermediate_edges(split_edges, admixture_edges)
}
\description{
For plotting the entire admixture graph, just nodes representing population
splits are not enough. We also need nodes (population states) which are not
explicitly simulated, but we need them to plot admixture edges appropriately.
}
