% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{graph}
\alias{graph}
\title{Plot admixture graph based on given model configuration}
\usage{
graph(model)
}
\arguments{
\item{model}{Compiled \code{spannr_model} model object}
}
\description{
Plot admixture graph based on given model configuration
}
