% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{graph}
\alias{graph}
\title{Plot admixture graph based on given model configuration}
\usage{
graph(populations, admixtures, removals = TRUE)
}
\arguments{
\item{populations}{List of \code{spammr_pop} objects}

\item{admixtures}{List of admixture events created by the
\code{admixture} function}

\item{removals}{Plot population removals?}
}
\description{
Plot admixture graph based on given model configuration
}
