% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{init_env}
\alias{init_env}
\title{Activate the Python environment of slendr}
\usage{
init_env(uv = FALSE, quiet = FALSE)
}
\arguments{
\item{uv}{Should an ephemeral Python environment be created via uv (instead
of activating a permanent virtual environment created via \code{setup_env})?}

\item{quiet}{Should informative messages be printed to the console? Default
is \code{FALSE}.}
}
\value{
No return value, called for side effects
}
\description{
This function activates a dedicated slendr Python environment set up via
\code{setup_env}, or creates an ephemeral Python environment with uv.
}
