% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{init_env}
\alias{init_env}
\title{Activate slendr's own dedicated Python environment}
\usage{
init_env(uv = FALSE, quiet = FALSE)
}
\arguments{
\item{uv}{Should an ephemeral Python environment be created via uv (instead
of activating a permanent virtual environment created via \code{setup_env})?}

\item{quiet}{Should informative messages be printed to the console? Default
is \code{FALSE}.}
}
\value{
No return value, called for side effects
}
\description{
This function attempts to activate a dedicated slendr Miniconda Python
environment previously set up via \code{setup_env}.
}
