% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{intersect_features}
\alias{intersect_features}
\title{Take a list of all population regions and intersect them with the
set of underlying world map}
\usage{
intersect_features(pop)
}
\description{
Take a list of all population regions and intersect them with the
set of underlying world map
}
