% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{migrate}
\alias{migrate}
\title{Move population to a new location in a given amount of time}
\usage{
migrate(region, towards, duration, snapshots = 5)
}
\description{
Move population to a new location in a given amount of time
}
