% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{migrate}
\alias{migrate}
\title{Move population to a new location in a given amount of time}
\usage{
migrate(pop, trajectory, duration, snapshots)
}
\arguments{
\item{pop}{Spatial population object of 'spammr_pop' class}

\item{trajectory}{List of two-dimensional vectors \link{(longitude, latitude)}
specifying the trajectory of the population movement}

\item{duration}{Duration of the population movement}

\item{snapshots}{Number of time slices to split the movement into}
}
\description{
Move population to a new location in a given amount of time
}
