% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{msprime}
\alias{msprime}
\title{Run a slendr model in msprime}
\usage{
msprime(
  model,
  sequence_length,
  recombination_rate,
  output = file.path(model$path, "output_msprime.trees"),
  sampling = NULL,
  verbose = TRUE,
  seed = NULL,
  debug = FALSE
)
}
\arguments{
\item{model}{Model object created by the \code{compile} function}

\item{sequence_length}{Total length of the simulated sequence (in base-pairs)}

\item{recombination_rate}{Recombination rate of the simulated sequence (in
recombinations per basepair per generation)}

\item{output}{A shared prefix path of output files that will be generated
by the model (by default, all files will share a prefix \code{"output"}
and will be placed in the model directory)}

\item{sampling}{A data frame of times at which a given number of individuals
should be remembered in the tree-sequence (see \code{sampling} for a
function that can generate the sampling schedule in the correct format). If
missing, only individuals present at the end of the simulation will be
recorded in the tree-sequence output file.}

\item{verbose}{Write the SLiM output log to the console (default
\code{FALSE})?}

\item{seed}{Random seed (if missing, SLiM's own seed will be used)}

\item{debug}{Print a msprime model debugging summary?}
}
\description{
This function will execute a built-in msprime script and run a compiled
slendr demographic model.
}
