% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{overlap}
\alias{overlap}
\title{Generate the overlap of two \code{spannr_region} objects}
\usage{
overlap(x, y, name = NULL)
}
\arguments{
\item{x}{Object of the class \code{spannr_region}}

\item{y}{Object of the class \code{spannr_region}}

\item{name}{Name of the resulting geographic region}
}
\value{
Object of the class \code{spannr_region}
}
\description{
Generate the overlap of two \code{spannr_region} objects
}
