% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.slendr}
\alias{plot.slendr}
\title{Plot \code{slendr} geographic features on a map}
\usage{
\method{plot}{slendr}(
  ...,
  pop_facets = TRUE,
  time_facets = FALSE,
  intersect = TRUE,
  graticules = "original",
  title = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{...}{Objects of classes \code{slendr_map},
\code{slendr_region}, or \code{slendr_pop}}

\item{pop_facets}{Plot populations in individual panels?}

\item{time_facets}{Plot time snapshots in individual panels?}

\item{intersect}{Intersect the population boundaries against
landscape and other geographic boundaries (default TRUE)?}

\item{graticules}{Plot graticules in the original Coordinate
Reference System (such as longitude-latitude), or in the internal
CRS (such as meters)?}

\item{title}{Plot title}

\item{nrow, ncol}{Number of columns or rows in the facet plot}
}
\description{
Plots objects of the three \code{slendr} spatial classes
(\code{slendr_map}, \code{slendr_region}, and \code{slendr_pop}).
}
