% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot.spammr}
\alias{plot.spammr}
\title{Plot spatio-temporal population distributions}
\usage{
\method{plot}{spammr}(..., facets = TRUE, rendering = TRUE, geo_graticules = TRUE)
}
\arguments{
\item{...}{Population/geographic region objects of the 'spammr'
class}

\item{facets}{Plot populations in individual panels?}

\item{rendering}{Render the population boundaries against landscape
and other geographic boundaries?}

\item{geo_graticules}{Plot axies with lon/lat graticules?}
}
\description{
Plot the spatio-temporal distributions of populations and
geographic regions across a map.
}
\details{
If only geographic regions are given, they are colored. If both
them and populations are given, only populations are specified.
}
