% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spammr.R
\name{plot.spammr}
\alias{plot.spammr}
\title{Plot spatio-temporal population distributions}
\usage{
\method{plot}{spammr}(
  ...,
  snapshots = NULL,
  facets = TRUE,
  intersect = TRUE,
  geo_graticules = TRUE,
  title = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{...}{Population/geographic region objects of the 'spammr'
class}

\item{snapshots}{Which time points to plot?}

\item{facets}{Plot populations in individual panels?}

\item{intersect}{intersect the population boundaries against landscape
and other geographic boundaries?}

\item{geo_graticules}{Plot axies with lon/lat graticules?}

\item{title}{Plot title}

\item{nrow, ncol}{Number of columns or rows in the facet plot}
}
\description{
Plot the spatio-temporal distributions of populations and
geographic regions across a map.
}
\details{
If only geographic regions are given, they are colored. If both
them and populations are given, only populations are specified.
}
