% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot.spamr}
\alias{plot.spamr}
\title{Plot spatio-temporal population distributions}
\usage{
\method{plot}{spamr}(..., snapshots = F)
}
\arguments{
\item{...}{Population range objects}

\item{snapshots}{Plot time snapshots in individual panels?}
}
\description{
Plot spatio-temporal population distributions
}
