% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_spatial.R
\name{plot_ancestors}
\alias{plot_ancestors}
\title{Plot locations of ancestors of given sample on a map}
\usage{
plot_ancestors(x, individual, connect = TRUE)
}
\description{
Plot locations of ancestors of given sample on a map
}
