% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ancestry}
\alias{plot_ancestry}
\title{Plot simulated ancestry proportions}
\usage{
plot_ancestry(model)
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}
}
\description{
Plot simulated ancestry proportions
}
