% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_model}
\alias{plot_model}
\title{Plot demographic history encoded in a slendr model}
\usage{
plot_model(model, sizes = TRUE)
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}

\item{sizes}{Should population size changes be visualized?}
}
\value{
Plot object of the class \code{ggplot}
}
\description{
Plot demographic history encoded in a slendr model
}
