% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_model}
\alias{plot_model}
\title{Plot demographic history encoded in a slendr model}
\usage{
plot_model(model, sizes = TRUE, proportions = FALSE, log = FALSE)
}
\arguments{
\item{model}{Compiled \code{slendr_model} model object}

\item{sizes}{Should population size changes be visualized?}

\item{proportions}{Should gene flow proportions be visualized (\code{FALSE}
by default to prevent cluttering and overplotting)}

\item{log}{Should the y-axis be plotted on a log scale? Useful for models
over very long time-scales.}
}
\value{
A ggplot2 object with the visualized slendr model
}
\description{
Plot demographic history encoded in a slendr model
}
\examples{
\dontshow{check_dependencies(python = TRUE) # make sure dependencies are present }
# load an example model with an already simulated tree sequence and plot
# it as an "admixture graph"
model <- read_example("introgression")

plot_model(model, log = TRUE)
}
