% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interface.R
\name{population}
\alias{population}
\title{Define a population range}
\usage{
population(
  name,
  parent,
  N,
  time = NULL,
  world = NULL,
  center = NULL,
  radius = NULL,
  coords = NULL,
  region = NULL,
  remove = NULL,
  intersect = TRUE
)
}
\arguments{
\item{name}{Name of the population}

\item{parent}{Parent population object or "ancestor" character scalar}

\item{N}{Number of individuals at the time of split}

\item{time}{Time of the population appearance}

\item{world}{Object of the type \code{sf} which defines the world}

\item{center}{Vector of two elements defining a center of a circular range}

\item{radius}{Scalar defining a radius of a range in kilometers}

\item{coords}{List of vector pairs, defining corners of the range}

\item{region}{Geographic region of the class \code{spammr_region}}

\item{remove}{Time at which the population should be removed}

\item{intersect}{Intersect the population's boundaries with landscape
features?}
}
\value{
Object of the \code{spammr_pop} (and \code{sf}) class
}
\description{
Define a population range
}
