% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ranges.R
\name{population}
\alias{population}
\title{Define a population range}
\usage{
population(
  name,
  time,
  world,
  center = NULL,
  radius = NULL,
  coords = NULL,
  region = NULL
)
}
\arguments{
\item{name}{Name of the population}

\item{time}{Time of the population appearance}

\item{world}{Object of the type 'sf' which defines the world}

\item{center}{Vector of two elements defining a center of a circular range}

\item{radius}{Scalar defining a radius of a range in kilometers}

\item{coords}{List of vector pairs, defining corners of the range}
}
\description{
Define a population range
}
