% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{population}
\alias{population}
\title{Define a population and its original spatial range}
\usage{
population(
  name,
  time,
  N,
  parent = "ancestor",
  map = FALSE,
  center = NULL,
  radius = NULL,
  polygon = NULL,
  remove = NULL,
  intersect = TRUE,
  competition_dist = NA,
  mate_dist = NA,
  dispersal_dist = NA,
  aquatic = FALSE
)
}
\arguments{
\item{name}{Name of the population}

\item{time}{Time of the population's first appearance}

\item{N}{Number of individuals at the time of first appearance}

\item{parent}{Parent population object or "ancestor" (indicating that the
population does not have an ancestor, and that it is the first population
in its "lineage")}

\item{map}{Object of the type \code{slendr_map} which defines the world
context (created using the \code{world} function). If the value
\code{FALSE} is provided, a non-spatial model will be run.}

\item{center}{Two-dimensional vector specifying the center of the circular
range}

\item{radius}{Radius of the circular range}

\item{polygon}{List of vector pairs, defining corners of the polygon range or
a geographic region of the class \code{slendr_region} from which the
polygon coordinates will be extracted (see the \code{region() function})}

\item{remove}{Time at which the population should be removed}

\item{intersect}{Intersect the population's boundaries with landscape
features?}

\item{competition_dist, mate_dist}{Maximum spatial competition and mating
choice distance}

\item{dispersal_dist}{Standard deviation of the normal distribution of the
distance that offspring disperses from its parent}

\item{aquatic}{Is the species aquatic (\code{FALSE} by default, i.e.
terrestrial species)?}
}
\value{
Object of the class \code{slendr_pop}
}
\description{
Defines a spatial range of a population and its most important parameters.
}
\details{
There are three ways to specify a spatial boundary: i) circular range
specified using a center coordinate and a radisu, ii) polygon specified as a
list of two-dimensional vector coordinates, iii) polygon as in ii), but
defined (and named) using the \code{region} function.
}
