% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{population}
\alias{population}
\title{Define a population and its original spatial range}
\usage{
population(
  name,
  time = NULL,
  N,
  parent,
  map = NULL,
  center = NULL,
  radius = NULL,
  coords = NULL,
  region = NULL,
  remove = NULL,
  intersect = TRUE
)
}
\arguments{
\item{name}{Name of the population}

\item{time}{Time of the population's first appearance}

\item{N}{Number of individuals at the time of first appearance}

\item{parent}{Parent population object or "ancestor" (indicating
that the population does not have an ancestor, and that it is the
first population in its "lineage")}

\item{map}{Object of the type \code{spannr_map} which defines the
world context (created using the \code{world} function)}

\item{center}{Two-dimensional vector specifying the center of the
circular range}

\item{radius}{Radius of the circular range}

\item{coords}{List of vector pairs, defining corners of the polygon
range (see also the \code{region} argument)}

\item{region}{Geographic region of the class \code{spannr_region}}

\item{remove}{Time at which the population should be removed}

\item{intersect}{Intersect the population's boundaries with
landscape features?}
}
\value{
Object of the class \code{spannr_pop}
}
\description{
Defines a spatial range of a population and its most important
parameters.
}
\details{
There are three ways to specify a spatial boundary: i) circular
range specified using a center coordinate and a radisu, ii) polygon
specified as a list of two-dimensional vector coordinates, iii)
polygon as in ii), but defined (and named) using the \code{region}
function.
}
