% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{population}
\alias{population}
\title{Define a population and its original spatial range}
\usage{
population(
  name,
  time,
  N,
  parent = "ancestor",
  map = FALSE,
  center = NULL,
  radius = NULL,
  polygon = NULL,
  remove = NULL,
  intersect = TRUE,
  competition_dist = NA,
  mate_dist = NA,
  dispersal_dist = NA,
  dispersal_fun = NULL,
  aquatic = FALSE
)
}
\arguments{
\item{name}{Name of the population}

\item{time}{Time of the population's first appearance}

\item{N}{Number of individuals at the time of first appearance}

\item{parent}{Parent population object or "ancestor" (indicating that the
population does not have an ancestor, and that it is the first population
in its "lineage")}

\item{map}{Object of the type \code{slendr_map} which defines the world
context (created using the \code{world} function). If the value
\code{FALSE} is provided, a non-spatial model will be run.}

\item{center}{Two-dimensional vector specifying the center of the circular
range}

\item{radius}{Radius of the circular range}

\item{polygon}{List of vector pairs, defining corners of the polygon range or
a geographic region of the class \code{slendr_region} from which the
polygon coordinates will be extracted (see the \code{region() function})}

\item{remove}{Time at which the population should be removed}

\item{intersect}{Intersect the population's boundaries with landscape
features?}

\item{competition_dist, mate_dist}{Maximum spatial competition and mating
choice distance}

\item{dispersal_dist}{Standard deviation of the normal distribution of the
distance that offspring disperses from its parent}

\item{dispersal_fun}{Distribution function governing the dispersal of
offspring. One of "normal", "uniform", "cauchy", or "exponential".}

\item{aquatic}{Is the species aquatic (\code{FALSE} by default, i.e.
terrestrial species)?}
}
\value{
Object of the class \code{slendr_pop}
}
\description{
Defines a spatial range of a population and its most important parameters.
}
\details{
There are three ways to specify a spatial boundary: i) circular range
specified using a center coordinate and a radisu, ii) polygon specified as a
list of two-dimensional vector coordinates, iii) polygon as in ii), but
defined (and named) using the \code{region} function.
}
\examples{
# create abstract landscape object of a given dimension
map <- world(xrange = c(0, 1000), yrange = c(0, 1000), landscape = "blank")

# maps can be plotted by the generic function `plot`
plot(map)

# create a circular population with the center of a population boundary at
# [600, 700] and a radius of 100 distance units, 1000 individuals at time 1
# occupying a map just specified
pop1 <- population("pop1", N = 1000, time = 1,
                   map = map, center = c(600, 700), radius = 100)

# create another population occupying a polygon range, splitting from pop1
# at a given time point (note that specifying a map is not necessary because
# it is "inherited" from the parent)
pop2 <- population("pop2", N = 100, time = 50, parent = pop1,
                   polygon = list(c(100, 100), c(500, 200),
                                  c(500, 400), c(100, 400)))

# check the position of both populations interactively by plotting their
# ranges on a map
plot(pop1, pop2)
}
