% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.slendr}
\alias{print.slendr}
\title{Print a summary of a \code{slendr} object}
\usage{
\method{print}{slendr}(x, ...)
}
\arguments{
\item{x}{Object of a class \code{slendr}}
}
\description{
Prints a short summary of any \code{slendr} object.
}
\details{
All spatial objects in the slendr package are internally
represented by a Simple Features (sf) object. This fact is hidden
in most circumstances this, as the goal of the slendr package is to
provide functionality at a much higher level (population
boundaries, geographic regions, instead of individual polygons and
other "low-level" geometric objects). The logical argument
\code{sf} allows the user to inspect the underlying object
structure. Similarly, by default, \code{print} does not print all
spatial map information to reduce clutter. If many individual
spatial maps for a population are present, a full table of spatial
snapshots can be printed using the \code{full} argument.
}
