% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spannr.R
\name{print.spannr}
\alias{print.spannr}
\title{Print a summary of a \code{spannr} object}
\usage{
\method{print}{spannr}(x, sf = FALSE, full = FALSE)
}
\arguments{
\item{x}{Object of a class \code{spannr}}

\item{sf}{Print the low-level 'sf' object instead?}

\item{full}{Print the complete table of spatial snapshots?}
}
\description{
Prints a short summary of any \code{spannr} object.
}
\details{
All spatial objects in the spannr package are internally
represented by a Simple Features (sf) object. This fact is hidden
in most circumstances this, as the goal of the spannr package is to
provide functionality at a much higher level (population
boundaries, geographic regions, instead of individual polygons and
other "low-level" geometric objects). The logical argument
\code{sf} allows the user to inspect the underlying object
structure. Similarly, by default, \code{print} does not print all
spatial map information to reduce clutter. If many individual
spatial maps for a population are present, a full table of spatial
snapshots can be printed using the \code{full} argument.
}
