% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rasterize}
\alias{rasterize}
\title{Render the population boundary to a black-and-white rasterized
spatial map}
\usage{
rasterize(..., rendering = TRUE)
}
\arguments{
\item{...}{Spatial population objects of the 'spammr_pop' class}

\item{rendering}{Render the population boundaries against landscape
and other geographic boundaries?}
}
\description{
Render the population boundary to a black-and-white rasterized
spatial map
}
