% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{rasterize}
\alias{rasterize}
\title{Rasterize the vector form of a population spatial boundary}
\usage{
rasterize(x, pixel)
}
\arguments{
\item{x}{Object of the 'spammr_pop' class}

\item{pixel}{Two-dimensional numeric vector specifying the
absolute dimension of the raster map}
}
\description{
Rasterize the vector form of a population spatial boundary
}
