% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{rasterize}
\alias{rasterize}
\title{Rasterize the vector form of a population spatial boundary}
\usage{
rasterize(x, resolution)
}
\arguments{
\item{x}{Object of the 'spannr_pop' class}

\item{resolution}{How many kilometers per pixel?}
}
\description{
Rasterize the vector form of a population spatial boundary
}
