% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{read}
\alias{read}
\title{Read a previously serialized model configuration}
\usage{
read(dir)
}
\arguments{
\item{dir}{Directory with all required configuration files}
}
\description{
Reads all configuration tables and other model data from a location
where it was previously compiled to by the \code{compile} function.
}
