% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{region}
\alias{region}
\title{Define a geographic region}
\usage{
region(name, map, center, radius, polygon)
}
\arguments{
\item{name}{Name of the geographic region}

\item{map}{Object of the type \code{sf} which defines the map}

\item{center}{Two-dimensional vector specifying the center of the
circular range}

\item{radius}{Radius of the circular range}

\item{polygon}{List of vector pairs, defining corners of the
polygon range or a geographic region of the class
\code{spannr_region} from which the polygon coordinates will be
extracted (see the \code{region() function})}
}
\value{
Object of the class \code{spannr_region}
}
\description{
Creates a geographic region (a polygon) on a given map and gives it
a name. This can be used to define objects which can be reused in
multiple places in a spannr script (such as \code{region} arguments
of \code{population}) without having to repeatedly define polygon
coordinates.
}
