% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{region}
\alias{region}
\title{Define a geographic region}
\usage{
region(name, world, coords)
}
\arguments{
\item{name}{Name of the geographic region}

\item{world}{Object of the type 'sf' which defines the world}

\item{coords}{List of vector pairs, defining corners of the range}
}
\description{
Define a geographic region
}
