% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{render}
\alias{render}
\title{Render population boundaries to black-and-white spatial maps}
\usage{
render(pops, resolution)
}
\description{
Render population boundaries to black-and-white spatial maps
}
