% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{render}
\alias{render}
\title{Take a list of all population regions and intersect them with the
set of underlying world map}
\usage{
render(pop)
}
\arguments{
\item{pop}{Spatial population object of the 'spammr_pop' class}
}
\description{
Take a list of all population regions and intersect them with the
set of underlying world map
}
