% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{resize}
\alias{resize}
\title{Resize the population size}
\usage{
resize(pop, N, how, time, end = NULL)
}
\arguments{
\item{pop}{Object of the class \code{slendr_pop}}

\item{N}{Population size after the change}

\item{how}{How to change the population size (options are \code{"step"},
\code{"exponential"}, \code{"linear"})}

\item{time}{Time of the population size change}

\item{end}{End of the population size change period (used for
exponential change events)}
}
\description{
Resize the population size
}
