% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{resize}
\alias{resize}
\title{Resize the population size}
\usage{
resize(pop, N, how, time, end = NULL)
}
\arguments{
\item{pop}{Object of the class \code{slendr_pop}}

\item{N}{Population size after the change}

\item{how}{How to change the population size (options are \code{"step"} or
\code{"exponential"})}

\item{time}{Time of the population size change}

\item{end}{End of the population size change period (used for exponential
change events)}
}
\description{
Resizes the population starting from the current value of \code{N}
individuals to the specified value
}
\details{
In the case of exponential size change, if the final \code{N} is larger than
the current size, the population will be exponentially growing over the
specified time period until it reaches \code{N} individuals. If \code{N} is
smaller, the population will shrink exponentially.
}
