% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{resize}
\alias{resize}
\title{Resize the population size}
\usage{
resize(pop, N, how = "step", time = NULL, start = NULL, end = NULL)
}
\arguments{
\item{pop}{Object of the class \code{slendr_pop}}

\item{N}{Final population size}

\item{how}{How to change the population size (options are "step",
"exponential", "linear")}

\item{time}{Time of the population size change}

\item{start, end}{Time-window for the population size change (used for
exponential or linear change)}
}
\description{
Resize the population size
}
