% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{run}
\alias{run}
\title{Open the compiled spatial model in SLiM}
\usage{
run(
  model,
  burnin,
  sim_length,
  seq_length,
  recomb_rate,
  max_distance,
  max_spread,
  save_locations = FALSE,
  track_ancestry = FALSE,
  gui = TRUE,
  verbose = FALSE,
  include = NULL,
  output_dir = model$config$directory
)
}
\arguments{
\item{model}{Compiled model object}

\item{burnin}{Length of the burnin (in model's time units, i.e. years)}

\item{sim_length}{Total length of the simulation (in model's time units, i.e.
years)}

\item{seq_length}{Total length of the simulated sequence in base-pairs}

\item{recomb_rate}{Recombination rate of the simulated sequence}

\item{max_distance, max_spread}{Maximum values for spatial interaction and
mating choice (max_distance) and maximum distance for offspring from its
parents (max_spread)}

\item{save_locations}{Save location of each individual throughout the
simulation?}

\item{track_ancestry}{Track ancestry proportion dynamics in all populations
throughout the simulations (default FALSE)? If a non-zero integer is
provided, ancestry will be tracked using the number number of neutral
ancestry markers equal to this number.}

\item{gui}{Run in SLiM GUI instead of command-line? (default)}

\item{include}{Vector of paths to custom SLiM scripts which should be
combined with the backend SLiM code}

\item{output_dir}{Where to put potential output files?}

\item{gen_time}{Generation time (in model's time units, i.e. years)}
}
\description{
When run, the compiled SLiM script will save the location of each individual
that ever lived, and will also record a sample of ancient individuals from
each population in a tree sequence data structure which will be saved for all
"present-day" individuals at the end of the simulation. This obviously does
not make sense for all potential uses and the exact specification of output
formats will be changed at some point soon.
}
