% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{run}
\alias{run}
\title{Open the SLiM backend script in the SLiM gui}
\usage{
run(
  model_dir,
  gen_time,
  burnin,
  sim_length,
  interaction,
  spread,
  seq_length,
  recomb_rate
)
}
\arguments{
\item{gen_time}{Conversion rate from model time units into generations}

\item{burnin}{Length of the burnin (in model time units)}

\item{sim_length}{Total length of the simulation (in model time units)}

\item{interaction}{Spatial interaction/mate choice distance parameter}

\item{spread}{Sigma parameter of the offspring spread normal distribution}

\item{seq_length}{Length of the simulated sequence}
}
\description{
Open the SLiM backend script in the SLiM gui
}
