% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_class}
\alias{set_class}
\title{Set slendr classes (or fix their priorities if already present)}
\usage{
set_class(x, type)
}
\description{
Set slendr classes (or fix their priorities if already present)
}
