% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{set_class}
\alias{set_class}
\title{Set spammr classes (or fix their priorities if already present)}
\usage{
set_class(x, type)
}
\arguments{
\item{x}{Object of the 'spammr' class}

\item{type}{Character scalar with the 'spammr' subtype name}
}
\description{
Set spammr classes (or fix their priorities if already present)
}
