% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/spammr.R
\name{set_class}
\alias{set_class}
\title{Set spammr classes (or fix their priorities if already present)}
\usage{
set_class(x, type)

set_class(x, type)
}
\description{
Set spammr classes (or fix their priorities if already present)

Set spammr classes (or fix their priorities if already present)
}
