% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{setup_env}
\alias{setup_env}
\title{Setup a dedicated Python virtual environment for slendr}
\usage{
setup_env(quiet = FALSE, agree = FALSE, pip = NULL)
}
\arguments{
\item{quiet}{Should informative messages be printed to the console? Default
is \code{FALSE}.}

\item{agree}{Automatically agree to all questions?}

\item{pip}{Should pip be used instead of conda for installing slendr's Python
dependencies? Note that this will still use the conda distribution to
install Python itself, but will change the repository from which slendr
will install its Python dependencies. Unless explicitly set to \code{TRUE},
Python dependencies will be installed from conda repositories by default,
expect for the case of osx-arm64 Mac architecture, for which conda
dependencies are broken.}
}
\value{
No return value, called for side effects
}
\description{
This function will automatically download a Python miniconda distribution
dedicated to an R-Python interface. It will also create a slendr-specific
Python environment with all the required Python dependencies.
}
