% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{setup_env}
\alias{setup_env}
\title{Setup a dedicated Python virtual environment for slendr}
\usage{
setup_env(env = NULL)
}
\arguments{
\item{env}{Either a name of or a path to a Python virtual environment, or a
name of a conda environment}
}
\description{
The environment will be called 'automatic_slendr_python_env' and if missing when loading the
slendr package, it will be automatically created and population with required
Python modules.
}
