% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{setup_env}
\alias{setup_env}
\title{Setup a dedicated Python virtual environment for slendr}
\usage{
setup_env(env = NULL, quiet = FALSE)
}
\arguments{
\item{env}{Either a name of a conda environment, or a path to a standard
Python virtual environment (such as one created by \code{python3 -m venv
  <path to an environment>}) with necessary Python dependencies (msprime,
tskit, pyslim, and pandas). If \code{NULL} (the default), the user will be
offered to install and configure an entirely isolated Miniconda Python
environment just for slendr. If this environment is already present,
calling \code{setup_env()} without the \code{env} argument will simply
activate this environment.}

\item{quiet}{Should informative messages be printed to the console? Default
is \code{FALSE}.}
}
\description{
Setup a dedicated Python virtual environment for slendr
}
