% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{shrink}
\alias{shrink}
\title{Shrink the population range}
\usage{
shrink(pop, by, end, start, overlap = 0.8, snapshots = NULL, verbose = TRUE)
}
\arguments{
\item{pop}{Object of the class \code{slendr_pop}}

\item{by}{How many units of distance to shrink by?}

\item{start, end}{When does the boundary shrinking start/end?}

\item{overlap}{Minimum overlap between subsequent spatial boundaries}

\item{snapshots}{The number of intermediate snapshots (overrides the
\code{overlap} parameter)}
}
\value{
Object of the class \code{slendr_pop}
}
\description{
Shrinks the spatial population range by a specified distance in a given
time-window
}
