% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{slim}
\alias{slim}
\title{Run a slendr model as a SLiM script}
\usage{
slim(
  model,
  seq_length,
  recomb_rate,
  save_locations = FALSE,
  track_ancestry = FALSE,
  ts_recording = FALSE,
  sampling = NULL,
  method,
  verbose = FALSE,
  include = NULL,
  burnin = 0,
  seed = NULL,
  slim_path = NULL
)
}
\arguments{
\item{model}{Model object created by the \code{compile} function}

\item{seq_length}{Total length of the simulated sequence (in base-pairs)}

\item{recomb_rate}{Recombination rate of the simulated sequence (in
recombinations per basepair per generation)}

\item{save_locations}{Save location of each individual throughout the
simulation?}

\item{track_ancestry}{Track ancestry proportion dynamics in all populations
throughout the simulations (default FALSE)? If a non-zero integer is
provided, ancestry will be tracked using the number number of neutral
ancestry markers equal to this number.}

\item{ts_recording}{Turn on tree sequence recording during SLiM
initialization?}

\item{method}{How to run the script? ("gui" - open in SLiMgui, "batch" - run
on the command-line, "script" - simply return the script)}

\item{verbose}{Write the SLiM output log to the console? (default
\code{FALSE})}

\item{include}{Vector of paths to custom SLiM scripts which should be
combined with the backend SLiM code}

\item{burnin}{Length of the burnin (in model's time units, i.e. years)}

\item{seed}{Random seed (if missing, SLiM's own seed will be used)}

\item{slim_path}{Optional way to specify path to an appropriate SLiM binary}

\item{samples}{A data frame of times at which a given number of individuals
should be remembered in the tree-sequence (see \code{sampling} for a function
that can generate the sampling schedule in the correct format). If missing,
only individuals present at the end of the simulation will be recorded in
the tree-sequence output file.}
}
\description{
Generates a SLiM script from a built-in slendr template, substituting all
required parameters by values specified by the user. The user has an option
to either run the script in batch mode, open it in SLiMgui, or simply save
the generated SLiM script to a given location for later use.
}
