% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{spatial_range}
\alias{spatial_range}
\title{Define a range (simple geometry object) for a population or a
geographic region}
\usage{
spatial_range(world, center = NULL, radius = NULL, coords = NULL)

spatial_range(world, center = NULL, radius = NULL, coords = NULL)
}
\description{
Define a range (simple geometry object) for a population or a
geographic region

Define a range (simple geometry object) for a population or a
geographic region
}
