% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{subtract}
\alias{subtract}
\title{Generate the difference between two \code{spannr_region} objects}
\usage{
subtract(x, y, name)
}
\arguments{
\item{x}{Object of the class \code{spannr_region}}

\item{y}{Object of the class \code{spannr_region}}

\item{name}{Name of the resulting geographic region}
}
\value{
Object of the class \code{spannr_region}
}
\description{
Generate the difference between two \code{spannr_region} objects
}
