% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_ancestors}
\alias{ts_ancestors}
\title{Infer spatio-temporal ancestral history for given nodes/individuals}
\usage{
ts_ancestors(ts, x = NULL, verbose = FALSE)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{x}{Either a character vector with individual names, or an integer
vector of node IDs}

\item{verbose}{Report on the progress of ancestry path generation?}
}
\description{
Infer spatio-temporal ancestral history for given nodes/individuals
}
