% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_coalesced}
\alias{ts_coalesced}
\title{Check that all trees in the tree sequence are fully coalesced}
\usage{
ts_coalesced(ts, return_failed = FALSE)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{return_failed}{Report back which trees failed the coalescence
check?}
}
\value{
TRUE or FALSE value if \code{return_failed = FALSE}, otherwise a vector of
(tskit Python 0-based) indices of trees which failed the coalescence test
}
\description{
Check that all trees in the tree sequence are fully coalesced
}
