% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_draw}
\alias{ts_draw}
\title{Plot a graphical representation of a single tree}
\usage{
ts_draw(x, width = 1500, height = 500, labels = FALSE, ...)
}
\arguments{
\item{x}{A single tree extracted by \code{\link{ts_tree}}}

\item{width, height}{Pixel dimensions of the rendered bitmap}

\item{labels}{Label each node with the individual name?}

\item{...}{Keyword arguments to the tskit \code{draw_svg} function.}
}
\description{
This function first obtains an SVG representation of the tree by calling the
\code{draw_svg} method of tskit and renders it as a bitmap image in R. All of
the many optional keyword arguments of the \code{draw_svg} method can be
provided and will be automatically passed to the method behind the scenes.
}
