% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_genotypes}
\alias{ts_genotypes}
\title{Extract genotype table from the tree sequence}
\usage{
ts_genotypes(ts)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}
}
\value{
Data frame object of the class \code{tibble}
}
\description{
Extract genotype table from the tree sequence
}
