% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_ibd}
\alias{ts_ibd}
\title{Collect Identity-by-Descent (IBD) tracts}
\usage{
ts_ibd(
  ts,
  coordinates = FALSE,
  within = NULL,
  between = NULL,
  min_length = NULL,
  max_time = NULL
)
}
\arguments{
\item{coordinates}{Should coordinates of all detected IBD tracts be reported?
If \code{FALSE} (the default), only the total pairwise IBD sharing statistics
are reported. If \code{TRUE}, coordinates of each segment will be given.}
}
\description{
This function iterates over a tree sequence and returns IBD tracts between
pairs of nodes (i.e. chromosomes or parts of chromosomes).
}
\details{
Iternally, this function leverages the tskit \code{TreeSequence} method
\code{ibd_segments}. Also, note that the \code{ts_ibd} function always returns
a data frame of IBD tracts, it does not provide an option to iterate over
individual IBD segments as shown in the official tskit documentation
\url{https://tskit.dev/tskit/docs/stable/ibd.html}. In general, R handles
heavy iteration poorly, and this function does not attempt to serve as
a full wrapper to \code{ibd_segments}.
}
