% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_individuals}
\alias{ts_individuals}
\title{Extract the table of individuals}
\usage{
ts_individuals(ts, remembered = NULL, retained = NULL, alive = NULL)
}
\arguments{
\item{ts}{pyslim.SlimTreeSequence object}

\item{model}{\code{slendr_model} object of a model that generated the data}
}
\value{
Data frame with individual information from the input tree sequence
object
}
\description{
Extract the table of individuals
}
