% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_load}
\alias{ts_load}
\title{Load a tree-sequence file produced by a given model (optionally simplifying
only to remembered individuals)}
\usage{
ts_load(file, model, simplify = FALSE)
}
\arguments{
\item{file}{Path to a tree-sequence file generated by SLiM}

\item{model}{Object of the class \code{slendr_model}}

\item{simplify}{Should the tree-sequence be simplified down to only
remembered (i.e. "sampled", in \textit{slendr} parlance) individuals?}
}
\value{
SlimTreeSequence object
}
\description{
Load a tree-sequence file produced by a given model (optionally simplifying
only to remembered individuals)
}
