% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_load}
\alias{ts_load}
\title{Load a tree-sequence file produced by a given model}
\usage{
ts_load(
  file,
  model,
  recapitate = FALSE,
  simplify = FALSE,
  recombination_rate = NULL,
  Ne = NULL,
  random_seed = NULL
)
}
\arguments{
\item{file}{Path to a tree-sequence file generated by SLiM}

\item{model}{Object of the class \code{slendr_model}}

\item{recapitate}{Should the tree-sequence be recapitated?}

\item{simplify}{Should the tree-sequence be simplified down to only
remembered (i.e. "sampled", in slendr parlance) individuals?}

\item{recombination_rate, Ne}{Arguments passed to \code{ts_recapitate}}

\item{random_seed}{Random seed passed to pyslim's \code{recapitate} method}
}
\value{
SlimTreeSequence object
}
\description{
This function loads a tree-sequence file simulated from a given slendr model.
Optionally, the tree-sequence can be recapitated and simplified.
}
\details{
The loading, recapitation and simplification is performed using the Python
module pyslim which serves as a link between tree-sequences generated by SLiM
and the tskit module for manipulation of tree-sequence data. All of these
steps have been modelled after the official pyslim tutorial and documentation
available at: \url{https://pyslim.readthedocs.io/en/latest/tutorial.html}.

The recapitation and simplification steps can also be performed individually
using the functions \code{ts_recapitate} and \code{ts_simplify}.
}
