% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_mutate}
\alias{ts_mutate}
\title{Add mutations to the given tree sequence}
\usage{
ts_mutate(ts, mutation_rate, random_seed = NULL, keep_existing = TRUE)
}
\arguments{
\item{ts}{Object of the type \code{pyslim.SlimTreeSequence}}

\item{mutation_rate}{Mutation rate used by msprime to simulate mutations}

\item{random_seed}{Random seed passed to msprime's \code{mutate} method}

\item{keep_existing}{Keep existing mutations?}
}
\value{
\code{pyslim.SlimTreeSequence} object of the class \code{slendr_ts}
}
\description{
Add mutations to the given tree sequence
}
\seealso{
\link{\code{\link{ts_data}}} for extracting useful information about
individuals, nodes, coalescent times and geospatial locations of nodes on a
map
}
