% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_names}
\alias{ts_names}
\title{Extract names of individuals of interest in a tree sequence produced
by a slendr model}
\usage{
ts_names(x, split = NULL, schedule = NULL)
}
\arguments{
\item{x}{Either a tree sequence object of the class \code{slendr_ts}, or
a compiled slendr model of the class \code{slendr_model}}

\item{split}{Should sample names in the tree sequence be split by a column
(a population or time column)? Default is \code{NULL} and all names of
samples will be returned as a single character vector. If set to "pop" or
"time", a list of character vectors will be returned, one vector for each
unique "pop" or "time" grouping.}

\item{schedule}{If samples from a model (not yet a tree sequence) are to be
obtained, this argument can be used to restrict the set of recorded
individuals}
}
\value{
A vector of character sample names. If \code{split} is specified,
a list of such vectors is returned, one element of the list per population
or sampling time.
}
\description{
Extract names of individuals of interest in a tree sequence produced
by a slendr model
}
