% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_nodes}
\alias{ts_nodes}
\alias{ts_edges}
\alias{ts_model}
\title{Get the table of individuals, nodes, or edges or a model object}
\usage{
ts_nodes(ts)

ts_edges(ts)

ts_model(ts)
}
\arguments{
\item{ts}{pyslim.SlimTreeSequence object}
}
\value{
Data frame
}
\description{
Get the table of individuals, nodes, or edges or a model object
}
