% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_phylo}
\alias{ts_phylo}
\title{Convert a tree in the tree sequence to an object of the class \code{phylo}}
\usage{
ts_phylo(ts, i, mode = c("index", "position"), quiet = FALSE)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{mode}{The mode for the calculation ("sites" or "branch")}

\item{quiet}{Should ape's internal phylo validity test be printed out?}
}
\description{
Convert a tree in the tree sequence to an object of the class \code{phylo}
}
