% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_recapitate}
\alias{ts_recapitate}
\title{Recapitate the tree sequence}
\usage{
ts_recapitate(
  ts,
  recombination_rate,
  Ne,
  random_seed = NULL,
  assign_metadata = FALSE
)
}
\arguments{
\item{ts}{pyslim.SlimTreeSequence object loaded by \code{ts_load}}

\item{recombination_rate}{A constant value of the recombination rate}

\item{Ne}{Effective population size during the recapitation process}

\item{random_seed}{Random seed passed to pyslim's \code{recapitate} method}
}
\value{
@return pyslim.SlimTreeSequence object accessed via the reticulate package
}
\description{
Recapitate the tree sequence
}
