% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_recapitate}
\alias{ts_recapitate}
\title{Recapitate the tree sequence}
\usage{
ts_recapitate(
  ts,
  recombination_rate,
  Ne,
  spatial = TRUE,
  migration_matrix = NULL,
  random_seed = NULL
)
}
\arguments{
\item{ts}{Tree sequence object loaded by \code{ts_load}}

\item{recombination_rate}{A constant value of the recombination rate}

\item{Ne}{Effective population size during the recapitation process}

\item{spatial}{Should spatial information encoded in the tree sequence data
be converted to spatial R data structures? If FALSE, pixel-based
raster-dimensions will not be converted to the coordinate reference system
implied by the model. If TRUE (default), reprojection of coordinates will
be performed. If the model was non-spatial, the value of this parameter is
disregarded.}

\item{migration_matrix}{Migration matrix used for coalescence of ancient lineages
(passed to \code{ts_recapitate})}

\item{random_seed}{Random seed passed to pyslim's \code{recapitate} method}
}
\value{
Tree sequence object of the class \code{slendr_ts}
}
\description{
Recapitate the tree sequence
}
\examples{
\dontrun{
ts_rec <- ts_recapitate(ts, recombination_rate = 1e-8, Ne = 10000, random_seed = 42)
}

}
\seealso{
\code{\link{ts_data}} for extracting useful information about
individuals, nodes, coalescent times and geospatial locations of nodes on a
map
}
