% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_samples}
\alias{ts_samples}
\title{Extract names and times of individuals of interest in a tree sequence produced
by a slendr model}
\usage{
ts_samples(x)
}
\arguments{
\item{x}{Either a tree sequence object of the class \code{slendr_ts}, or
a compiled slendr model of the class \code{slendr_model}}
}
\value{
Table of individuals scheduled for sampling across space and time
}
\description{
Extracts a table of metadata information about individuals recorded in
a given simulated tree sequence or in a tree sequence which would be
produced by a given model
}
\details{
If a model is provided as \code{x}, the function will first simulate
a trivially small tree sequence using the built-in msprime engine.
}
\examples{
\dontshow{check_dependencies(python = TRUE, quit = TRUE) # dependencies must be present
}
init_env()

# load an example model with an already simulated tree sequence
slendr_ts <- system.file("extdata/models/introgression_slim.trees", package = "slendr")
model <- read_model(path = system.file("extdata/models/introgression", package = "slendr"))

# load the tree-sequence object from disk
ts <- ts_read(slendr_ts, model)

# extract the table of individuals scheduled for simulation and sampling
ts_samples(ts)
}
