% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_samples}
\alias{ts_samples}
\title{Extract information about permanently remembered individuals}
\usage{
ts_samples(ts)
}
\description{
Extract information about permanently remembered individuals
}
