% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_samples}
\alias{ts_samples}
\title{Extract names and times of individuals scheduled for sampling}
\usage{
ts_samples(ts)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}
}
\description{
Extract names and times of individuals scheduled for sampling
}
