% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_save}
\alias{ts_save}
\title{Write a tree sequence to a file}
\usage{
ts_save(ts, file)
}
\arguments{
\item{ts}{Tree sequence object loaded by \code{ts_read}}

\item{file}{File to which the tree sequence should be saved}
}
\description{
Deprecated function. Please use \code{ts_write} instead.
}
