% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_simplify}
\alias{ts_simplify}
\title{Simplify tree-sequence object down to only remembered (i.e. "sampled", in
\textit{slendr} parlance) individuals}
\usage{
ts_simplify(ts)
}
\arguments{
\item{ts}{SlimTreeSequence object}
}
\value{
@return SlimTreeSequence object accessed via the reticulate package
}
\description{
Simplify tree-sequence object down to only remembered (i.e. "sampled", in
\textit{slendr} parlance) individuals
}
