% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_tajima}
\alias{ts_tajima}
\title{Calculate Tajima's D for given sets of individuals}
\usage{
ts_tajima(ts, sample_sets, mode = c("site", "branch", "node"), windows = NULL)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{sample_sets}{A list (optionally a named list) of character vectors with
individual names (one vector per set). If a simple vector is provided, it
will be interpreted as \code{as.list(sample_sets)}, meaning that a given
statistic will be calculated for each individual separately.}

\item{mode}{The mode for the calculation ("sites" or "branch")}

\item{windows}{Coordinates of breakpoints between windows. The first
coordinate (0) and the last coordinate (equal to \code{ts$sequence_length})
are added automatically)}
}
\value{
For each set of individuals either a single Tajima's D value or a
vector of Tajima's D values (one for each window)
}
\description{
For a discussion on the difference between "site" and "branch" options of the
\code{mode} argument, please see the tskit documentation at
\url{https://tskit.dev/tskit/docs/stable/stats.html#sec-stats-mode}
}
\examples{
# the examples will only run when a dedicated Python environment is present
# (this can be created by calling `setup_env()`)
if (check_env(quiet = TRUE)) {

# load an example model with an already simulated tree sequence
model <- read_example("introgression")

ts <- ts_load(model, mutate = TRUE, mutation_rate = 1e-8)

# calculate Tajima's D for given sets of individuals in a tree sequence ts
ts_tajima(ts, list(afr = c("AFR_1", "AFR_2", "AFR_3", "AFR_4", "AFR_5"),
                   eur = c("EUR_1", "EUR_2", "EUR_3", "EUR_4", "EUR_5"),
                   nea = c("NEA_1", "NEA_2")))
}

}
