% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_tajima}
\alias{ts_tajima}
\title{Calculate Tajima's D for given sets of individuals}
\usage{
ts_tajima(ts, sample_sets, mode = c("site", "branch", "node"), windows = NULL)
}
\arguments{
\item{ts}{pyslim.SlimTreeSequence object}

\item{sample_sets}{A list (optionally a named list) of character vectors with
individual names (one vector per set)}

\item{mode}{The mode for the calculation ("sites" or "branch")}

\item{windows}{Coordinates of breakpoints between windows. The first
coordinate (0) and the last coordinate (equal to \code{ts$sequence_length})
are added automatically)}
}
\value{
For each set of individuals either a single Tajima's D value or a
vector of Tajima's D values (one for each window)
}
\description{
For a discussion on the difference between "site" and "branch" options of the
\code{mode} argument, please see the tskit documentation at
\url{https://tskit.dev/tskit/docs/stable/stats.html#sec-stats-mode}
}
