% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspop.R
\name{ts_tracts}
\alias{ts_tracts}
\title{Extract ancestry tracts
See \url{https://tspop.readthedocs.io/en/latest/basicusage.html} for more
details.}
\usage{
ts_tracts(ts, census, squashed = TRUE, source = NULL)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{census}{Census time (in slendr time units)}

\item{squashed}{Should ancestry tracts be squashed?}

\item{source}{Source population to filter ancestry tracts for}
}
\description{
Extract ancestry tracts
See \url{https://tspop.readthedocs.io/en/latest/basicusage.html} for more
details.
}
