% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequences.R
\name{ts_tree}
\alias{ts_tree}
\title{Get a tree from a given tree sequence}
\usage{
ts_tree(ts, i, mode = c("index", "position"), ...)
}
\arguments{
\item{ts}{Tree sequence object of the class \code{slendr_ts}}

\item{i}{Position of the tree in the tree sequence. If \code{mode = "index"},
an i-th tree will be returned (in one-based indexing), if \code{mode =
  "position"}, a tree covering an i-th base of the simulated genome will be
returned.}

\item{mode}{How should the \code{i} argument be interpreted? Either "index"
as an i-th tree in the sequence of genealogies, or "position" along the
simulated genome.}

\item{...}{Additional keyword arguments accepted by
\code{tskit.TreeSequence.at and tskit.TreeSequence.at_index} methods}
}
\value{
Object of the type tskit.trees.Tree
}
\description{
For more information about optional keyword arguments see tskit documentation:
\url{https://tskit.dev/tskit/docs/stable/python-api.html#the-treesequence-class}
}
\examples{
\dontrun{ # extract the first tree in the tree sequence
tree <- ts_tree(ts, i = 1)

# extract the tree at a position 100000bp in the tree sequence
tree <- ts_tree(ts, i = 100000, mode = "position")
}

}
