% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{world}
\alias{world}
\title{Define a world map for all spatial operations}
\usage{
world(xrange, yrange, landscape = "naturalearth", crs = NULL, ne_dir = NULL)
}
\arguments{
\item{xrange}{Two-dimensional vector specifying minimum and maximum
horizontal range ("longitude" if using real Earth cartographic
data)}

\item{yrange}{Two-dimensional vector specifying minimum and maximum
vertical range ("latitude" if using real Earth cartographic data)}

\item{landscape}{Either "blank" (for blank abstract geography),
"naturalearth" (for real Earth geography) or an object of the
class \code{sf} defining abstract geographic features of the
world}

\item{crs}{EPSG code of a coordinate reference system to use for
spatial operations. No CRS is assumed by default (\code{NULL}),
implying an abstract landscape not tied to any real-world
geographic region (when \code{landscape = "blank"} or when
\code{landscape} is a custom-defined geographic landscape), or
implying WGS-84 (EPSG 4326) coordinate system when a real Earth
landscape was defined (\code{landscape = "naturalearth"}).}

\item{ne_dir}{Path to the directory where Natural Earth data was
manually downloaded and unzipped from
\url{https://www.naturalearthdata.com/downloads/110m-physical-vectors}
(used only when \code{landscape = "naturalearth"})}
}
\value{
Object of the class \code{slendr_map}
}
\description{
Defines either an abstract geographic landscape (blank or
containing user-defined landscape) or using a real Earth
cartographic data from the Natural Earth project
(\url{https://www.naturalearthdata.com}).
}
