% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{write_model}
\alias{write_model}
\title{Write a compiled slendr model to disk}
\usage{
write_model(
  dir,
  populations,
  admix_table,
  map_table,
  split_table,
  generation_time,
  resolution,
  length
)
}
\description{
Write a compiled slendr model to disk
}
